% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_degree}
\alias{add_degree}
\title{add_degree}
\usage{
add_degree(x, add)
}
\arguments{
\item{x}{Add describtion}

\item{add}{Add describtion}
}
\value{
Add describtion
}
\description{
Add describtion
}
\details{
Add describtion
}
\keyword{export}
