% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_result.R
\name{animate_result}
\alias{animate_result}
\title{animate_result}
\usage{
animate_result(result, fill = "reef")
}
\arguments{
\item{result}{RasterBrick with environment created with \code{\link{setup_environment}}.}

\item{fill}{Character to specify which values of environmental data is used as fill.}

\item{...}{Arguments passed on to \code{gganimate::animate}.}
}
\value{
data frame
}
\description{
Wrapper functions to run model
}
\details{
Wrapper function to run model. Executes the following sub-processes (i) ...
(ii) ...

Parameters include ...
}
\examples{
\dontrun{

animate_result()

}

}
