% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrR-packages.R
\docType{package}
\name{arrR}
\alias{arrR}
\alias{arrR-package}
\title{arrR}
\description{
Individual-based simulation model to analyze how artificial reefs in combination
with a fish population influence seagrass growth.

For more information about how to use the model, please see \url{https://allgeier-lab.github.io/arrR/}.

For a detailed model description, see Esquivel, K.E., Hesselbarth, M.H.K., Allgeier, J.E.
Mechanistic support for increased primary production around artificial reefs. Manuscript
submitted for publication.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Allgeier-Lab/arrR}
  \item Report bugs at \url{https://github.com/Allgeier-Lab/arrR/issues}
}

}
\author{
\strong{Maintainer}: Maximillian H.K. Hesselbarth \email{mhk.hesselbarth@gmail.com} (\href{https://orcid.org/0000-0003-1125-9918}{ORCID})

Authors:
\itemize{
  \item Kenzo E. Esquivel \email{kesquivel@berkeley.edu} (\href{https://orcid.org/0000-0002-1326-289X}{ORCID})
  \item Katrina S. Munsterman \email{kmunster@umich.edu} (\href{https://orcid.org/0000-0002-9303-2049}{ORCID})
}

Other contributors:
\itemize{
  \item Samantha Iliff \email{siliff@umich.edu} (\href{https://orcid.org/0000-0002-3070-4565}{ORCID}) [illustrator]
  \item Jacob E. Allgeier \email{jeallg@umich.edu} (\href{https://orcid.org/0000-0002-9005-6432}{ORCID}) [research team head]
}

}
