% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arrR_parameters}
\alias{arrR_parameters}
\title{Default parameters}
\format{
List
}
\usage{
arrR_parameters
}
\description{
List with default parameters.
}
\references{
Citations for all parameters and starting values can be found in
Esquivel et al (2021). Mechanistic support for increased primary
production around artificial reefs. Manuscript in preparation.
}
\keyword{datasets}
