% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_dist_reefs.R
\name{int_calc_dist_reefs}
\alias{int_calc_dist_reefs}
\alias{calc_dist_reefs}
\title{calc_dist_reefs}
\usage{
int_calc_dist_reefs(fish_population, coords_reef)
}
\arguments{
\item{fish_population}{2-column matrix with coordinates of individual fish.}

\item{coords_reef}{2-column matrix with coordinates of AR.}
}
\value{
vector
}
\description{
Internal function
}
\details{
Internal function calculate distance to reef cells.
}
\keyword{internal}
