% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_monod.R
\name{calc_monod}
\alias{calc_monod}
\title{calc_monod}
\usage{
calc_monod(nutrients, v_max, k_m)
}
\arguments{
\item{nutrients}{Vector with amount of nutrients.}

\item{v_max}{Numeric with maximum uptake rate.}

\item{k_m}{Numeric with half-saturation rate.}
}
\value{
Vector
}
\description{
Internal function
}
\details{
Internal function to calculate nutrient uptake (Formula 3.7).
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6

Lee, K.-S., Dunton, K.H., 1999. Inorganic nitrogen acquisition in the seagrass
Thalassia testudinum: Development of a whole-plant nitrogen budget.
Limnol. Oceanogr. 44, 1204–1215. https://doi.org/10.4319/lo.1999.44.5.1204
}
