% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nutr_uptake.R
\name{calc_nutr_uptake}
\alias{calc_nutr_uptake}
\title{calc_nutr_uptake}
\usage{
calc_nutr_uptake(nutrients, biomass, v_max, k_m, time_fac)
}
\arguments{
\item{nutrients}{Vector with nutrients pool values.}

\item{biomass}{Vector with biomass values.}

\item{v_max, k_m}{Parameters of Michaelis–Menten model.}

\item{time_fac}{Numeric that specifies time period one iteration corresponds to.}
}
\value{
vector
}
\description{
Internal function to calculate nutrient uptake
}
\details{
Calculates the nutrient uptake for a given nutrients concentration in the water
column and biomass. The total uptake is based on the Michaelis–Menten model which
describes the acutal uptake in relation to the total possible uptake based on
the nutrient concentration.
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6

Lee, K.-S., Dunton, K.H., 1999. Inorganic nitrogen acquisition in the seagrass
Thalassia testudinum: Development of a whole-plant nitrogen budget.
Limnol. Oceanogr. 44, 1204–1215. https://doi.org/10.4319/lo.1999.44.5.1204
}
