% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_size.R
\name{int_calc_size}
\alias{int_calc_size}
\alias{calc_size}
\title{calc_size}
\usage{
int_calc_size(starting_values, parameters)
}
\arguments{
\item{starting_values}{List with all starting value parameters.}

\item{parameters}{List with all model parameters.}
}
\value{
vector
}
\description{
Internal function
}
\details{
Internal function to calculate size and weight of individuals.
}
\keyword{internal}
