% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary.R
\name{calc_summary}
\alias{calc_summary}
\title{calc_summary}
\usage{
calc_summary(x, foo, what)
}
\arguments{
\item{x}{Seafloor_values or fish_population.}

\item{foo}{Summary function.}

\item{what}{Character.}
}
\value{
vector
}
\description{
Calculate sigmoid to distribute uptake.
}
\details{
Calculates foo summary function of object cols.
}
