% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_max_biomass.R
\name{check_max_biomass}
\alias{check_max_biomass}
\title{check_max_biomass}
\usage{
check_max_biomass(seafloor_values, biomass_max, gamma)
}
\arguments{
\item{seafloor_values}{Matrix with seafloor values.}

\item{biomass_max}{Vector with maximum biomass.}

\item{gamma}{Vector with conversion factor between biomass and nutrients.}
}
\value{
Matrix
}
\description{
Check if biomass is above maximum.
}
\details{
Checks if biomass is above maximum biomass and if so, adds difference to detritus pool.
For biomass_max and gamma the first element of the vectors must be the bg value
and the second element the ag value.
}
