% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameters.R
\name{check_parameters}
\alias{check_parameters}
\title{check_parameters}
\usage{
check_parameters(starting_values, parameters)
}
\arguments{
\item{starting_values}{List with all starting value parameters.}

\item{parameters}{List with all model parameters.}
}
\value{
vector
}
\description{
Check model parameters
}
\details{
The function checks if all required starting values and parameters are provided.
If values are missing, a vector with the name of the corresponding values will
be returned.
}
\examples{
starting_values <- system.file("extdata", "starting_values.csv", package = "coRal")
parameters <- system.file("extdata", "parameters.csv", package = "coRal")

starting_values <- read_parameters(file = starting_values, sep = ";")
parameters <- read_parameters(file = parameters, sep = ";")

check_parameters(starting_values = starting_values, parameters = parameters)

}
