% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_nutr.R
\name{convert_nutr}
\alias{convert_nutr}
\title{convert_nutr}
\usage{
convert_nutr(x, to)
}
\arguments{
\item{x}{Numeric value to convert}

\item{to}{Character specifying unit to convert to. Either "g" (gram) or "um" (micromole).}
}
\value{
numeric
}
\description{
Convert nutrients from g to umol (and vice versa)
}
\details{
Converts the amount of nutrients from gram to micromole or the other way around.
The conversion is based on the molecular mass of NH4.
}
