% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRal-packages.R
\docType{package}
\name{coRal}
\alias{coRal}
\alias{coRal-package}
\title{coRal}
\description{
More about what it does (maybe more than one line)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Allgeier-Lab/coRAL}
  \item Report bugs at \url{https://github.com/Allgeier-Lab/coRAL/issues}
}

}
\author{
\strong{Maintainer}: Maximillian H.K. Hesselbarth \email{mhk.hesselbarth@gmail.com} (\href{https://orcid.org/0000-0003-1125-9918}{ORCID})

Authors:
\itemize{
  \item Kenzo Esquivel \email{kesquivel@berkeley.edu}
}

Other contributors:
\itemize{
  \item Jacob E. Allgeier \email{jeallg@umich.edu} [research team head]
}

}
