% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_parameters}
\alias{default_parameters}
\title{Default parameters}
\format{
List
}
\usage{
default_parameters
}
\description{
List with default parameters.
}
\references{
Citations for all parameters and starting values can be found in Esquivel, K.E.,
Hesselbarth, M.H.K., Allgeier, J.E., In Press. Mechanistic support for increased
primary production around artificial reefs. Ecological Applications. v0.0
}
\keyword{datasets}
