% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_starting}
\alias{default_starting}
\title{Default starting values}
\format{
List
}
\usage{
default_starting
}
\description{
List with default starting values.
}
\references{
Citations for all parameters and starting values can be found in Esquivel, K.E., Hesselbarth, M.H.K.,
Allgeier, J.E., 2022. Mechanistic support for increased primary production around
artificial reefs. Ecological Applications e2617. <https://doi.org/10.1002/eap.2617>
}
\keyword{datasets}
