% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_dead_detritus.R
\name{distribute_dead_detritus}
\alias{distribute_dead_detritus}
\title{distribute_dead_detritus}
\usage{
distribute_dead_detritus(seafloor, detritus_death_decomp)
}
\arguments{
\item{seafloor}{Environment created with \code{\link{setup_seafloor}}.}

\item{detritus_death_decomp}{}
}
\value{
RasterBrick
}
\description{
Simulate seagrass.
}
\details{
Function to simulate processes of above ground and below ground seagrass.
}
