% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_result.R
\name{extract_result}
\alias{extract_result}
\title{extract_result}
\usage{
extract_result(result, extract)
}
\arguments{
\item{result}{List with results of model run.}

\item{extract}{Character specifying if "seafloor" or "fishpop" is extracted.}
}
\value{
data.frame
}
\description{
Extract result
}
\details{
Extract only seafloor or fishpop values as data.frame.
}
\examples{
# Add example code

}
