% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mdlrn.R
\name{filter_mdlrn}
\alias{filter_mdlrn}
\title{filter_mdlrn}
\usage{
filter_mdlrn(result, timestep = max(result$max_i))
}
\arguments{
\item{result}{mdl_rn object of simulation run.}

\item{timestep}{Vector with timesteps to select}
}
\value{
data.frame
}
\description{
Filter model run results
}
\details{
Function to filter results according to timestep
}
\examples{
# Add example code

}
