% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_result.R
\name{filter_result}
\alias{filter_result}
\title{filter_result}
\usage{
filter_result(result, timestep = max(result$max_i))
}
\arguments{
\item{result}{mdl_rn object of simulation run.}

\item{timestep}{Vector with timesteps to select}
}
\value{
data.frame
}
\description{
Filter result.
}
\details{
Function to filter results according to timestep
}
\examples{
# Add example code

}
