% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density.R
\name{get_density}
\alias{get_density}
\title{get_density}
\usage{
get_density(result, normalize = FALSE, verbose = TRUE)
}
\arguments{
\item{result}{mdl_rn object.}

\item{normalize}{Logical if TRUE, count is divided by time steps.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
data.frame
}
\description{
Get density of fish individuals.
}
\details{
Calculates the fish density for each seafloor cells. Thus, the total count of
fish abundance within a raster cell is divided by the maximum time step.

Please keep in mind that if not each time step was saved during \code{\link{run_simulation}},
the returned density might not be the "true" density, because some occurrences
might be missed.
}
\examples{
\dontrun{
get_density(result = result_rand)
}

}
