% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_limits.R
\name{get_limits}
\alias{get_limits}
\title{get_limits}
\usage{
get_limits(result, timestep = result[[1]]$max_i)
}
\arguments{
\item{result}{List with results of model run.}

\item{timestep}{Integer to specify which timestep is plotted.}
}
\description{
Get density of fish
}
\details{
Returns list with minimum and maximum values of several result runs for better
plotting.
}
\examples{
\dontrun{

example code

}

}
\seealso{
\code{\link{run_simulation}}
}
