% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbors.R
\name{get_neighbors}
\alias{get_neighbors}
\title{get_neighbors}
\usage{
get_neighbors(x, direction = 8, cpp = FALSE)
}
\arguments{
\item{x}{RasterLayer.}

\item{direction}{Integer specifying if 4 or 8 neighborhood rule should be applied.}
}
\value{
matrix
}
\description{
Get ID of neighboring cells.
}
\details{
Internal function to get matrix with cell IDs of all neighboring cells.
The \code{direction} arguments allows to specifiy which cells are considered to be
neighbors (direction = 4: "Rook's case"; direction = 8: "Queen's case").
}
\examples{
\dontrun{
get_neighbors(seafloor)
}

}
