% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_production.R
\name{get_production}
\alias{get_production}
\title{get_production}
\usage{
get_production(result, lag = TRUE)
}
\arguments{
\item{result}{mdl_rn object of simulation run.}

\item{lag}{Logical if cumulative or difference to previous timestep should be returned.}
}
\value{
data.frame
}
\description{
Summarize production of model run.
}
\details{
Function to summarize the ag and bg production of each timestep. The \code{lag}
argument allows to not return the cumulative value, but the difference to the previous
timestep.
}
\examples{
\dontrun{
get_production(result_rand)
}

}
