% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stable_values.R
\name{get_stable_values}
\alias{get_stable_values}
\title{get_stable_values}
\usage{
get_stable_values(starting_values, parameters)
}
\arguments{
\item{starting_values}{List with all starting value parameters.}

\item{parameters}{List with all model parameters.}
}
\value{
list
}
\description{
Get starting values for stable growth.
}
\details{
Returns a list with starting values which allow stable seagrass growth if no
fish individuals are present.
}
\examples{
get_stable_values(starting_values = default_starting_values,
parameters = default_parameters)

}
