% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starting_values.R
\name{get_starting_values}
\alias{get_starting_values}
\title{get_starting_values}
\usage{
get_starting_values(seafloor_values, fishpop_values)
}
\arguments{
\item{seafloor_values}{Matrix with seafloor values.}

\item{fishpop_values}{Matrix population values.}
}
\value{
list
}
\description{
Get mean starting values of model run.
}
\details{
Get mean starting values. Includes values for i) bg_biomass, ii) ag_biomass,
iii) nutrients_pool, iv) detritus_pool, v) pop_n, and vi) pop_mean_size.
The function is needed because some of the values can include stochasticity
and can differ from the provided starting values.
}
\examples{
\dontrun{
get_starting_values(seafloor_values = seafloor_values, fishpop_values = fishpop_values)
}

}
\keyword{internal}
