% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{insert_random_phrase}
\alias{insert_random_phrase}
\title{Insert a random phrase into your vector of sentences}
\usage{
insert_random_phrase(split_sentences, phrases = pirate_phrases)
}
\arguments{
\item{split_sentences}{vector of individual sentences}

\item{phrases}{tibble of phrases with a phrase column
Default: pirate_phrases}
}
\value{
vector of sentences with a random phrase inserted at some position
}
\description{
inserts random phrase
}
\examples{
insert_random_phrase(c("Okay here's sentence 1.", "Here's sentence two"))
}
