% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_as_data_table_ras.R
\name{int_as_data_table_ras}
\alias{int_as_data_table_ras}
\title{as_data_table_ras}
\usage{
int_as_data_table_ras(
  x,
  xy = TRUE,
  in_memory = raster::canProcessInMemory(x),
  ...
)
}
\arguments{
\item{x}{Raster* object.}

\item{...}{Additional arguments passed to \code{`as.data.frame`} or
\code{`getValues`}.}
}
\value{
data.table
}
\description{
Convert raster to data.table
}
\details{
Internal function to convert Raster* to data.table.
}
\references{
Based on code by Etienne B. Racine with contribution of Lyndon Estes taken
from https://gist.github.com/etiennebr/9515738.
}
\keyword{internal}
