% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/int_calc_dist_reefs.R
\name{rcpp_calc_dist_reef}
\alias{rcpp_calc_dist_reef}
\alias{int_calc_dist_reef}
\title{int_calc_dist_reef}
\usage{
rcpp_calc_dist_reef(fish_population, coords_reef)

int_calc_dist_reef(fish_population, coords_reef)
}
\arguments{
\item{fish_population}{2-column matrix with coordinates of individual fish.}

\item{coords_reef}{2-column matrix with coordinates of AR.}
}
\value{
matrix

vector
}
\description{
Internal function

Internal function
}
\details{
Internal function calculate distance to reef cells.

Internal function calculate distance to reef cells.
}
\keyword{internal}
