% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_nutr_uptake.R
\name{int_calc_nutr_uptake}
\alias{int_calc_nutr_uptake}
\title{calc_nutr_uptake}
\usage{
int_calc_nutr_uptake(nutrients, biomass, v_max, k_m, time_fac)
}
\arguments{
\item{nutrients}{Add info about parameter.}

\item{biomass}{Add info about parameter.}

\item{v_max, k_m}{Add info about parameter.}

\item{time_fac}{Add info about parameter.}
}
\value{
vector
}
\description{
Internal function to calculate nutrient uptake
}
\details{
Add details
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6

Lee, K.-S., Dunton, K.H., 1999. Inorganic nitrogen acquisition in the seagrass
Thalassia testudinum: Development of a whole-plant nitrogen budget.
Limnol. Oceanogr. 44, 1204–1215. https://doi.org/10.4319/lo.1999.44.5.1204
}
\keyword{internal}
