% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_get_neighbors.R
\name{int_get_neighbors}
\alias{int_get_neighbors}
\title{int_get_neighbors}
\usage{
int_get_neighbors(x, direction = 4, torus = FALSE)
}
\arguments{
\item{x}{RasterLayer.}

\item{direction}{Integer specifying if 4 or 8 neighborhood rule should be applied.}

\item{torus}{If TRUE RasterLayer is treated as torus.}
}
\value{
matrix
}
\description{
Simulate diffusion.
}
\details{
Return matrix with cell ids of neighboring cells.
}
\keyword{internal}
