% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_grow_seagrass.R
\name{int_grow_seagrass}
\alias{int_grow_seagrass}
\title{grow_seagrass}
\usage{
int_grow_seagrass(nutrients, gamma, detritus_ratio)
}
\arguments{
\item{nutrients}{Add info about parameter.}

\item{gamma}{Add info about parameter.}

\item{detritus_ratio}{Add info about parameter.}
}
\value{
RasterBrick
}
\description{
Internal function to simulate above ground seagrass
}
\details{
Internal function to simulate below ground seagrass processes. This
includes i) ... ii) ...

#' @references
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6
}
\keyword{internal}
