% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_foo.R
\name{int_calc_foo}
\alias{int_calc_foo}
\alias{int_int_calc_foo}
\title{calc_foo}
\usage{
int_calc_foo(x, foo, what)
}
\arguments{
\item{x}{Seafloor_values or fish_population.}

\item{foo}{Summary function.}

\item{what}{Character.}
}
\value{
vector
}
\description{
Calculate sigmoid to distribute uptake.
}
\details{
Calculates foo summary function of object cols.
}
\keyword{internal}
