% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_sigmoid.R
\name{int_calc_sigmoid}
\alias{int_calc_sigmoid}
\alias{int_int_calc_sigmoid}
\title{calc_sigmoid}
\usage{
int_calc_sigmoid(x, log_slope)
}
\arguments{
\item{x}{Numeric value to convert.}

\item{what}{Either 'above' or 'below'.}
}
\value{
vector
}
\description{
Calculate sigmoid to distribute uptake.
}
\details{
Calculates sigmoid that redistributes total uptake depending on x.
}
\keyword{internal}
