% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_uptake.R
\name{int_calc_uptake}
\alias{int_calc_uptake}
\alias{int_int_calc_uptake}
\title{int_calc_uptake}
\usage{
int_calc_uptake(nutrients, biomass, v_max, k_m)
}
\arguments{
\item{nutrients}{Vecto with amount of wc nutrients.}

\item{biomass}{Vector with biomass.}

\item{v_max}{Numeric with maximum uptake rate.}

\item{k_m}{Numeric with half-saturation rate.}
}
\value{
Vector
}
\description{
Internal function
}
\details{
Internal function to calculate nutrient uptake (Formula 3.7).
}
\keyword{internal}
