% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_convert_dry.R
\name{int_convert_dry}
\alias{int_convert_dry}
\alias{int_int_convert_dry}
\title{convert_dry}
\usage{
int_convert_dry(x, what)
}
\arguments{
\item{x}{Numeric value to convert.}

\item{what}{Either 'above' or 'below'.}
}
\value{
numeric
}
\description{
Convert wet biomass to dry biomass
}
\details{
Converts above ground wet biomass to dry biomass.
}
\keyword{internal}
