% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_rebirth.R
\name{int_rebirth}
\alias{int_rebirth}
\title{rebirth}
\usage{
int_rebirth(
  fish_population,
  fish_population_track,
  a_grunt,
  b_grunt,
  n_body,
  parameters,
  detritus_pool,
  detritus_dead,
  reason
)
}
\arguments{
\item{fish_population, fish_population_track}{Data frame population created
with \code{\link{setup_fish_population}}.}

\item{detritus_pool, detritus_dead}{Vector with detritus values at location of
individual.}

\item{reason}{Character specifying reason of death ('consumption' or 'background').}
}
\value{
data.frame
}
\description{
Internal function
}
\details{
Internal function to create new individual if mortality takes place
}
\keyword{internal}
