% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_seagrass_accel.R
\name{int_seagrass_accel}
\alias{int_seagrass_accel}
\title{seagrass_accel}
\usage{
int_seagrass_accel(
  biomass_dry_bg,
  biomass_dry_ag,
  biomass_max_bg,
  nutrients,
  nutrients_thres,
  sigmoid_slope,
  v_max,
  k_max,
  gamma,
  slough_ratio,
  slough_detritus_ratio
)
}
\arguments{
\item{nutrients}{Add info about parameter.}

\item{nutrients_thres}{Add info about parameter.}

\item{sigmoid_slope}{Add info about parameter.}

\item{v_max}{Add info about parameter.}

\item{k_max}{Add info about parameter.}

\item{gamma}{Add info about parameter.}

\item{slough_ratio}{Add info about parameter.}

\item{slough_detritus_ratio}{Add info about parameter.}

\item{biomass_dry}{Add info about parameter.}

\item{max_biomass}{Add info about parameter.}
}
\value{
RasterBrick
}
\description{
Internal function to simulate accelerated seagrass
}
\details{
Internal function to simulate above ground seagrass processes. This
includes i) ... ii) ...
}
\keyword{internal}
