% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_seagrass_bg.R
\name{int_seagrass_bg}
\alias{int_seagrass_bg}
\title{seagrass_bg}
\usage{
int_seagrass_bg(
  biomass_dry,
  nutrients,
  sigmoid_slope,
  nutrients_thres_a,
  nutrients_thres_b,
  v_max,
  k_max,
  gamma,
  slough_ratio,
  slough_detritus_ratio,
  verbose
)
}
\arguments{
\item{biomass_dry}{Add info about parameter.}

\item{nutrients}{Add info about parameter.}

\item{sigmoid_slope}{Add info about parameter.}

\item{nutrients_thres_a}{Add info about parameter.}

\item{nutrients_thres_b}{Add info about parameter.}

\item{v_max}{Add info about parameter.}

\item{k_max}{Add info about parameter.}

\item{gamma}{Add info about parameter.}

\item{slough_ratio}{Add info about parameter.}

\item{slough_detritus_ratio}{Add info about parameter.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
RasterBrick
}
\description{
Internal function to simulate above ground seagrass
}
\details{
Internal function to simulate below ground seagrass processes. This
includes i) ... ii) ...
}
\keyword{internal}
