% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_seagrass_growth.R
\name{int_seagrass_growth}
\alias{int_seagrass_growth}
\title{seagrass_growth}
\usage{
int_seagrass_growth(
  biomass_growth,
  biomass_reduction,
  nutrients,
  growth_fraction,
  reduction_fraction,
  gamma,
  slough_ratio,
  slough_detritus_ratio,
  reduction
)
}
\arguments{
\item{biomass_growth}{Add info about parameter.}

\item{biomass_reduction}{Add info about parameter.}

\item{nutrients}{Add info about parameter.}

\item{gamma}{Add info about parameter.}

\item{slough_ratio}{Add info about parameter.}

\item{slough_detritus_ratio}{Add info about parameter.}
}
\value{
RasterBrick
}
\description{
Internal function to simulate above ground seagrass
}
\details{
Internal function to simulate below ground seagrass processes. This
includes i) ... ii) ...
}
\keyword{internal}
