% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_setup_envir_values.R
\name{int_setup_envir_values}
\alias{int_setup_envir_values}
\title{setup_envir_values}
\usage{
int_setup_envir_values(
  seafloor,
  ag_biomass,
  bg_biomass,
  wc_nutrients,
  detritus_fraction
)
}
\arguments{
\item{seafloor}{Raster* object.}

\item{ag_biomass, bg_biomass, wc_nutrients}{Numeric with starting values.}

\item{detritus_fraction}{Numeric with parameter for detritus fraction.}
}
\value{
vector
}
\description{
Internal function
}
\details{
Internal function to set biomass cell values of non-ARs cells.
}
\keyword{internal}
