% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_update_i_envir.R
\name{int_update_i_envir}
\alias{int_update_i_envir}
\title{update_i_envir}
\usage{
int_update_i_envir(environment, environment_dt, increase = 1)
}
\arguments{
\item{environment}{Environment created with \code{\link{setup_environment}}.}

\item{environment_dt}{Data table of environment.}

\item{increase}{Numeric how much i is increased.}
}
\value{
data.table
}
\description{
Update the time step counter of the environment.
}
\details{
Internal function to increase the time step i after each simulated time step.
}
\keyword{internal}
