% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_update_i_pop.R
\name{int_update_i_pop}
\alias{int_update_i_pop}
\title{update_i_pop}
\usage{
int_update_i_pop(population, increase = 1)
}
\arguments{
\item{population}{Data table with population data.}

\item{increase}{Numeric how much i is increased.}
}
\value{
data.table
}
\description{
Update the time step counter.
}
\details{
Internal function to increase the time step i after each simulated time step.
}
\keyword{internal}
