% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlrn_to_raster.R
\name{mdlrn_to_raster}
\alias{mdlrn_to_raster}
\title{mdlrn_to_raster}
\usage{
mdlrn_to_raster(mdl_rn, verbose = TRUE, ...)
}
\arguments{
\item{mdl_rn}{mdl_rn object.}

\item{verbose}{If TRUE, progress reports are printed.}

\item{...}{Additional arguments passed on to \code{\link{rast}}.}
}
\value{
SpatRaster
}
\description{
Convert mdl_rn object to SpatRaster.
}
\details{
Function to convert the seafloor from a previous \code{mdl_rn} object to
a \code{SpatRaster}. Thus, the created seafloor will have the final values of the
provided \code{mdl_rn} object as cell values.
}
\examples{
\dontrun{
mdlrn_to_raster(mdl_rn = result_rand)
}

}
