% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pirate_lookup}
\alias{pirate_lookup}
\title{Pirate Speak Lookup Table}
\format{
A data frame with 182 rows and 2 variables:
\describe{
\item{\code{orig}}{character english word}
\item{\code{new}}{character pirate word}
}
}
\source{
\url{https://github.com/ntoll/arrr/blob/master/arrr.py}

\url{https://github.com/mikewesthad/pirate-speak/blob/master/lib/pirate-speak.js}
}
\usage{
pirate_lookup
}
\description{
This lookup table maps English to pirate speak.
}
\keyword{datasets}
