% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pirate_phrases}
\alias{pirate_phrases}
\title{Pirate Phrases}
\format{
A data frame with 17 rows and 1 variables:
\describe{
\item{\code{phrase}}{character phrase in pirate speak}
}
}
\source{
\url{https://github.com/ntoll/arrr/blob/master/arrr.py}
}
\usage{
pirate_phrases
}
\description{
Collection of phrases that a pirate might randomly say.
}
\examples{
pirate_phrases
}
\keyword{datasets}
