% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mdl_rn.R
\name{plot.mdl_rn}
\alias{plot.mdl_rn}
\title{plot.mdl_rn}
\usage{
\method{plot}{mdl_rn}(
  x,
  what = "seafloor",
  summarize = FALSE,
  timestep = x$max_i,
  limits = NULL,
  burn_in = TRUE,
  normalize = FALSE,
  base_size = 10,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{mdl_rn object of simulation run.}

\item{what}{Character specifying what to plot.}

\item{summarize}{Character to specify which values of environmental data is used as fill.}

\item{timestep}{Integer to specify which timestep is plotted.}

\item{limits}{Named list with vectors with min and maximum value of values.}

\item{burn_in}{If TRUE, line to indicate burn-in time is plotted.}

\item{normalize}{Logical if TRUE count is divided by timesteps.}

\item{base_size}{Numeric to specify base font size.}

\item{verbose}{If TRUE, progress reports are printed.}

\item{...}{Not used.}
}
\description{
Plotting method for \code{mdl_rn} object.
}
\details{
Plotting method for model run results simulated with \code{\link{run_simulation}}.
The \code{what} arguments allows to either plot the seafloor values or fishpop values.
For both, it is possible to either plot a spatial raster at a certain timestep
\code{summarize = FALSE} or a summary of the values over all saved  timestep \code{summarize = TRUE}.
If a spatial raster is plotted, the limits can be set identical. For more information see
\code{\link{get_limits}}.
}
\examples{
\dontrun{
plot(result_rand)
}

}
