% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_production.R
\name{plot_production}
\alias{plot_production}
\title{plot_production}
\usage{
plot_production(result, summarize = FALSE, lag = TRUE)
}
\arguments{
\item{result}{mdl_rn object.}

\item{summarize}{Logical if values over time should be plotted.}

\item{lag}{Logical if TRUE, cumulative or difference to previous time step should be returned.}
}
\description{
Plotting method.
}
\details{
Plotting method for \code{mdl_rn} object simulated with \code{\link{run_simulation}}
for bg and ag biomass production.
}
\examples{
\dontrun{
plot_production(result = result_rand)
}

}
