% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_threshold.R
\name{plot_threshold}
\alias{plot_threshold}
\title{plot_threshold}
\usage{
plot_threshold(mdl_rn, timestep = NULL, base_size = 10)
}
\arguments{
\item{mdl_rn}{\code{mdl_rn} object of \code{\{link{simulation run}}.}

\item{timestep}{Integer with timestep to plot. If \code{NULL}, all timesteps are plotted.}

\item{base_size}{Numeric to specify base font size.}
}
\description{
Plotting method for cells above threshold
}
\details{
.
Plots all belowground cells in which biomass is above \code{seagrass_thres}
}
\examples{
# Add example code

}
