% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mdl_rn.R
\name{print.mdl_rn}
\alias{print.mdl_rn}
\title{print.mdl_rn}
\usage{
\method{print}{mdl_rn}(x, digits = 3, ...)
}
\arguments{
\item{x}{mdl_rn object.}

\item{digits}{Numeric of decimal places (round).}

\item{...}{Arguments passed to cat.}
}
\description{
Printing method.
}
\details{
Printing method for \code{mdl_rn} object simulated with \code{\link{run_simulation}}.
Includes information about the model run and resulting seafloor and fishpop values.
}
\examples{
\dontrun{
print(result_rand)
}


}
