% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_add_input}
\alias{rcpp_add_input}
\title{rcpp_add_input}
\usage{
rcpp_add_input(seafloor, nutr_input, timestep)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{nutr_input}{Vector with amount of nutrient input each timestep.}

\item{timestep}{Integer with current timestep.}
}
\value{
void
}
\description{
Rcpp add nutrient input
}
\details{
Rcpp implementation to add nutrient input to raster cells.
}
