% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_allocation_ratio}
\alias{rcpp_allocation_ratio}
\title{rcpp_allocation_ratio}
\usage{
rcpp_allocation_ratio(biomass, biomass_min, biomass_max, threshold, slope)
}
\arguments{
\item{biomass}{Numeric with biomass.}

\item{biomass_min, biomass_max}{Numeric with minum and maximum of biomass.}

\item{threshold, slope}{Numeric with function parameters.}
}
\value{
void
}
\description{
Rcpp allocation ratio
}
\details{
Rcpp implementation of the allocation ratio
}
