% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_dist_fish}
\alias{rcpp_calc_dist_fish}
\title{rcpp_calc_dist_fish}
\usage{
rcpp_calc_dist_fish(fish_population, coords_reef)
}
\arguments{
\item{fish_population}{2-column matrix with coordinates of individual fish.}

\item{coords_reef}{2-column matrix with coordinates of AR.}
}
\value{
matrix
}
\description{
Internal rcpp function
}
\details{
Internal function calculate distance to reef cells.
}
\keyword{export}
\keyword{internal}
