% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_dist_reef}
\alias{rcpp_calc_dist_reef}
\title{rcpp_calc_dist_reef}
\usage{
rcpp_calc_dist_reef(seafloor, coords_reef, extent, torus = FALSE)
}
\arguments{
\item{seafloor}{Matrix.}

\item{coords_reef}{2-column matrix with coordinates of AR.}

\item{extent}{Vector with dimension in x and y direction.}

\item{torus}{If TRUE the distance will be calculated using a torus.}
}
\value{
matrix
}
\description{
Rcpp calc dist reef
}
\details{
Internal function to calculate distance to reef cells.
}
\keyword{export}
