% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_growth}
\alias{rcpp_calc_growth}
\title{rcpp_calc_growth}
\usage{
rcpp_calc_growth(
  fishpop,
  seafloor,
  growth_values,
  pop_n_body,
  pop_max_reserves,
  min_per_i
)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{seafloor}{Matrix with seafloor values.}

\item{growth_values}{Matrix with growth values.}

\item{pop_n_body, pop_max_reserves, min_per_i}{Numeric with parameters.}
}
\value{
Matrix
}
\description{
Rcpp calc growth
}
\details{
\code{Rcpp} implementation of to calculate growth.
}
