% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_mineralization}
\alias{rcpp_calc_mineralization}
\title{rcpp_calc_mineralization}
\usage{
rcpp_calc_mineralization(
  seafloor,
  detritus_dead_decomp,
  detritus_mineralization
)
}
\arguments{
\item{seafloor}{Add describtion}

\item{detritus_dead_decomp, detritus_mineralization}{seafloor Add describtion}
}
\value{
Add describtion
}
\description{
Add describtion
}
\details{
Add describtion
}
\keyword{export}
