% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_mortality_background}
\alias{rcpp_calc_mortality_background}
\title{rcpp_calc_mortality_background}
\usage{
rcpp_calc_mortality_background(
  fishpop,
  fishpop_track,
  seafloor,
  fish_id,
  cell_id,
  pop_linf,
  pop_n_body,
  pop_want_reserves
)
}
\arguments{
\item{fishpop, fishpop_track}{Matrix with fishpop and starting fishpop values.}

\item{seafloor}{Matrix with seafloor values.}

\item{fish_id, cell_id}{Vector with id of fish and corresponding cell ids.}

\item{pop_linf, pop_n_body, pop_want_reserves}{Numeric with parameters.}
}
\value{
void
}
\description{
Rcpp background mortality
}
\details{
Rcpp implementation to create new individual after background mortality event.
}
