% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_reincarnate}
\alias{rcpp_reincarnate}
\alias{rcpp_calc_mortality_consump}
\title{rcpp_reincarnate}
\usage{
rcpp_reincarnate(
  fishpop,
  fishpop_track,
  seafloor,
  fish_id,
  cell_id,
  pop_linf,
  pop_n_body,
  pop_want_reserves,
  reason
)
}
\arguments{
\item{fishpop, fishpop_track}{Matrix with fishpop and starting fishpop values.}

\item{seafloor}{Matrix with seafloor values.}

\item{fish_id, cell_id}{Vector with id of fish and corresponding cell ids.}

\item{pop_linf, pop_n_body, pop_want_reserves}{Numeric with parameters.}

\item{reason}{String with reason of reincarnation.}
}
\value{
void
}
\description{
Rcpp reincarnate
}
\details{
Rcpp implementation to create new individual after mortality event.
}
