% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_nutr_uptake}
\alias{rcpp_calc_nutr_uptake}
\title{rcpp_calc_nutr_uptake}
\usage{
rcpp_calc_nutr_uptake(nutrients, biomass, v_max, k_m, time_frac)
}
\arguments{
\item{nutrients, biomass}{Add describtion}

\item{v_max, k_m, time_frac}{Add describtion}
}
\value{
Add describtion
}
\description{
Add describtion
}
\details{
Add describtion
}
\keyword{export}
