% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_respiration}
\alias{rcpp_calc_respiration}
\title{rcpp_calc_respiration}
\usage{
rcpp_calc_respiration(
  fishpop,
  resp_intercept,
  resp_slope,
  resp_temp_low,
  resp_temp_max,
  resp_temp_optm,
  water_temp,
  min_per_i
)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{resp_intercept, resp_slope}{Numeric with parameters.}

\item{resp_temp_low, resp_temp_max, resp_temp_optm}{Numeric with parameters.}

\item{water_temp, min_per_i, }{Numeric with parameters.}
}
\value{
void
}
\description{
Rcpp calculate respration
}
\details{
Rcpp implementation to calculate respiration of fish individuals.
}
