% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_cell_from_xy}
\alias{rcpp_cell_from_xy}
\title{rcpp_cell_from_xy}
\usage{
rcpp_cell_from_xy(x, y, extent, dimensions, rcpp)
}
\arguments{
\item{x, y}{Numeric with x,y coordinates.}

\item{extent}{Vector with extent (xmin, xmax, ymin, ymax).}

\item{dimensions}{Vector with number or rows and cols}

\item{rcpp}{Logical if TRUE Rcpp index is returned.}
}
\value{
int
}
\description{
Rcpp get cell from xy
}
\details{
Get cell ID from xy coordinates. Allows only one coordinate pair at a time.
If \code{rcpp = TRUE} indexing starts at 0 in accordance with C++.
}
\references{
Code adapted from Robert J. Hijmans (2020). raster: Geographic Data Analysis
and Modeling. R package version 3.4-5. <https://CRAN.R-project.org/package=raster>
}
\keyword{internal}
