% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_check_max_biomass}
\alias{rcpp_check_max_biomass}
\title{rcpp_check_max_biomass}
\usage{
rcpp_check_max_biomass(
  bg_biomass,
  ag_biomass,
  detritus_pool,
  bg_biomass_max,
  ag_biomass_max
)
}
\arguments{
\item{bg_biomass, ag_biomass, detritus_pool}{Add describtion}

\item{bg_biomass_max, ag_biomass_max}{Add describtion}
}
\value{
Add describtion
}
\description{
Add describtion
}
\details{
Add describtion
}
\keyword{export}
