% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_closest_reef}
\alias{rcpp_closest_reef}
\title{rcpp_closest_reef}
\usage{
rcpp_closest_reef(x, y, coords_reef)
}
\arguments{
\item{coords_reef}{Matrix with ID and coords of reef cells.}

\item{coords_temp}{Vector with xy coords of current individual.}
}
\value{
vector
}
\description{
Rcpp get closest reef.
}
\details{
Get ID and distance to closet reef cell. The first element of the returning
vector is the ID, the second the distance in meter.
}
