% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_convert_nutr}
\alias{rcpp_convert_nutr}
\title{rcpp_convert_nutr}
\usage{
rcpp_convert_nutr(x, to)
}
\arguments{
\item{x}{Add describtion}

\item{to}{Add describtion}
}
\value{
Add describtion
}
\description{
Add describtion
}
\details{
Add describtion
}
\keyword{export}
