% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_diffuse_values}
\alias{rcpp_diffuse_values}
\title{rcpp_diffuse_values}
\usage{
rcpp_diffuse_values(
  seafloor_values,
  cell_adj,
  wc_diffusion,
  detritus_diffusion,
  detritus_dead_diffusion
)
}
\arguments{
\item{seafloor_values}{Matrix with seafloor values.}

\item{cell_adj}{Matrix with cell adjacencies.}

\item{wc_diffusion, detritus_diffusion, detritus_dead_diffusion}{Numeric with parameters.}
}
\value{
Matrix
}
\description{
Rcpp sample function
}
\details{
\code{Rcpp} implementation of to diffuse nutrients.
}
