% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_dist_reef}
\alias{rcpp_dist_reef}
\title{rcpp_dist_reef}
\usage{
rcpp_dist_reef(seafloor, coords_reef, extent, torus)
}
\arguments{
\item{seafloor}{Matrix with coordinates of cells.}

\item{coords_reef}{2-column Matrix with coordinates of AR.}

\item{extent}{Vector with dimension in x and y direction.}

\item{torus}{If TRUE the distance will be calculated using a torus.}
}
\value{
vector
}
\description{
Rcpp calc dist reef
}
\details{
Rcpp implementation to calculate distance to reef cells. Returns vector with
distance for each cell to reef
}
\keyword{export}
