% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_adjacencies}
\alias{rcpp_get_adjacencies}
\title{rcpp_get_adjacencies}
\usage{
rcpp_get_adjacencies(dimensions, torus)
}
\arguments{
\item{dimensions}{Vector with number or rows and cols}

\item{torus}{Logical if true, torus translation is used.}
}
\value{
matrix
}
\description{
Rcpp get adjacencies
}
\details{
Get matrix with cell IDs of all neighboring cells. Indices start with 0
according to C++ indexing.
}
\references{
Code adapted from Robert J. Hijmans (2020). raster: Geographic Data Analysis
and Modeling. R package version 3.4-5. <https://CRAN.R-project.org/package=raster>
}
\keyword{internal}
