% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_bearing}
\alias{rcpp_get_bearing}
\title{rcpp_get_bearing}
\usage{
rcpp_get_bearing(x1, y1, x2, y2)
}
\arguments{
\item{x1, y1}{Double with first xy coords pair.}

\item{x2, y2}{Double with second xy coords pair.}
}
\value{
double
}
\description{
Rcpp get bearing between two coordinate pairs.
}
\details{
Get bearing between (x1,y1) and (x2,y2).
}
