% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_neighbors}
\alias{rcpp_get_neighbors}
\title{rcpp_get_neighbors}
\usage{
rcpp_get_neighbors(id, n_cell, n_col, directions)
}
\arguments{
\item{id}{Integer with id of focal cell.}

\item{n_col, n_cell}{Integer with number of columns and cells.}

\item{directions}{Integer specifying neighborhood rule}
}
\value{
vector
}
\description{
Rcpp get id of neighboring cells
}
\details{
Returns a vector with cell ID of neighboring cells. The direction argument can be
set to \code{directions = 4} (Rook's case) or \code{directions = 8} (Queen's case).
The returning vector gives the ID in the following order: left, top, right, bottom or
left, top_left, top, top_right, right, bottom_right, bottom, bottom_left.
}
