% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_reef}
\alias{rcpp_get_reef}
\title{rcpp_get_reef}
\usage{
rcpp_get_reef(seafloor)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}
}
\value{
matrix
}
\description{
Rcpp get reef matrix
}
\details{
Get matrix with cell id and coordinates of reef cells.
}
