% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mineralization}
\alias{rcpp_mineralization}
\title{rcpp_mineralization}
\usage{
rcpp_mineralization(seafloor, detritus_fish_ratio, detritus_mineralization)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{detritus_fish_ratio, detritus_mineralization}{seafloor Numeric with parameters.}
}
\value{
void
}
\description{
Rcpp calculate mineralization
}
\details{
Rcpp implementation to calculate detritus miniralization.
}
