% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mortality}
\alias{rcpp_mortality}
\title{rcpp_mortality}
\usage{
rcpp_mortality(
  fishpop,
  fishpop_track,
  seafloor,
  pop_ldie,
  pop_n_body,
  pop_reserves_max,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop, fishpop_track}{Matrix with fishpop and starting fishpop values.}

\item{seafloor}{Matrix with seafloor values.}

\item{pop_ldie, }{pop_n_body,pop_reserves_max Numeric with parameters.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}
}
\value{
void
}
\description{
Rcpp simulate (background) mortality.
}
\details{
Function to simulate background mortality of fish individuals.
Fish automatically dies when \code{pop_mean_size} + 2.
If a individual dies, a new individual is created.
}
\keyword{internal}
