% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_behav}
\alias{rcpp_move_behav}
\title{rcpp_move_behav}
\usage{
rcpp_move_behav(
  fishpop,
  coords_reef,
  pop_thres_reserves,
  move_mean,
  move_var,
  move_reef,
  move_border,
  move_return,
  max_dist,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{coords_reef}{Matrix with coords of reef cells.}

\item{pop_thres_reserves}{Vector with threshold of pop_max_reserves to drain prior to foraging.}

\item{move_mean, move_var}{Double with mean movement parameter.}

\item{move_reef}{Double with mean movement distance when sheltering at reef.}

\item{move_border}{Double with movement distance that surrounds reef cell border.}

\item{move_return}{Double with mean movement distance when returning to reef.}

\item{max_dist}{Maximum distance an individual can move.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}
}
\value{
void
}
\description{
Rcpp move fish population
}
\details{
Rcpp implementation to move fish individuals depending on move distance and
heading value.
}
