% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_fishpop}
\alias{rcpp_move_fishpop}
\title{rcpp_move_fishpop}
\usage{
rcpp_move_fishpop(fishpop, move_dist, extent, pop_mean_move)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{move_dist}{Vector with move distance of fish individuals.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{pop_mean_move}{Numeric with parameter.}
}
\value{
void
}
\description{
Rcpp move fish population
}
\details{
Rcpp implementation to move fish individuals depending on move distance and
heading value.
}
