% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rcpp_nutr_input}
\alias{.rcpp_nutr_input}
\alias{rcpp_nutr_input}
\title{rcpp_nutr_input}
\usage{
.rcpp_nutr_input(seafloor, nutrients_input)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{nutrients_input}{Vector with amount of nutrient input each time step.}
}
\value{
void
}
\description{
Rcpp simulate nutrient input.
}
\details{
Simulate external nutrient input to the each cell. The \code{nutrients_input}
vector must have as many elements as \code{max_i} to add input each time step.

If no nutrients should be added, set all values of \code{nutrients_input} to
zero.
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. <https://doi.org/10.1007/978-94-011-2342-6>
}
\keyword{internal}
