% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_nutr_uptake}
\alias{rcpp_nutr_uptake}
\title{rcpp_nutr_uptake}
\usage{
rcpp_nutr_uptake(nutrients, biomass, v_max, k_m, time_frac)
}
\arguments{
\item{nutrients, biomass}{Numeric with nutrient and biomass amount of cell.}

\item{v_max, k_m, time_frac}{Numeric with parameters}
}
\value{
double
}
\description{
Rcpp nutrient uptake
}
\details{
Calculate nutrient uptake of cells.
}
\references{
Lee, K.-S., Dunton, K.H., 1999. Inorganic nitrogen acquisition in the seagrass
Thalassia testudinum: Development of a whole-plant nitrogen budget.
Limnol. Oceanogr. 44, 1204–1215.https://doi.org/10.4319/lo.1999.44.5.1204
}
