% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_quantile}
\alias{rcpp_quantile}
\title{rcpp_quantile}
\usage{
rcpp_quantile(x, q)
}
\arguments{
\item{x}{NumericVector with values.}

\item{q}{Double with quantile.}
}
\value{
double
}
\description{
Rcpp quantile
}
\details{
Returns q-th percentile of vector.
}
\references{
Rcpp quantile implementation from
<https://stackoverflow.com/questions/26786078/rcpp-quantile-implementation>
}
\keyword{internal}
