% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_remove_output}
\alias{rcpp_remove_output}
\title{rcpp_remove_output}
\usage{
rcpp_remove_output(seafloor, nutrients_output)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{nutrients_output}{Double with fraction removed from each cell.}
}
\value{
void
}
\description{
Rcpp remove nutrients
}
\details{
Rcpp implementation to remove nutrient from raster cells.
}
