% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_rlognorm}
\alias{rcpp_rlognorm}
\title{rcpp_rlognorm}
\usage{
rcpp_rlognorm(mean, sd)
}
\arguments{
\item{mean}{Double with mean.}

\item{sd}{Double with sd}
}
\value{
double
}
\description{
Create random number from log distribution
}
\details{
Get random number from log-norm distribution. Function uses log-transformed
values and a normal distribution internally.
}
\keyword{export}
