% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_runif}
\alias{rcpp_runif}
\title{rcpp_runif}
\usage{
rcpp_runif(min, max)
}
\arguments{
\item{min, max}{Double boundaries.}
}
\value{
double
}
\description{
Rcpp runif.
}
\details{
Draws random number from uniform distribution.
}
\references{
<https://www.cplusplus.com/reference/random/uniform_real_distribution/>
}
\keyword{internal}
