% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_seagrass_growth}
\alias{rcpp_seagrass_growth}
\title{rcpp_seagrass_growth}
\usage{
rcpp_seagrass_growth(
  seafloor,
  cells_reef,
  bg_v_max,
  bg_k_m,
  bg_gamma,
  ag_v_max,
  ag_k_m,
  ag_gamma,
  bg_biomass_max,
  bg_biomass_min,
  ag_biomass_max,
  ag_biomass_min,
  seagrass_thres,
  seagrass_slope,
  seagrass_slough,
  time_frac
)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{cells_reef}{Vector with id of reef cells.}

\item{bg_v_max, bg_k_m, bg_gamma, ag_v_max, ag_k_m, ag_gamma}{Numeric with uptake parameters.}

\item{bg_biomass_max, bg_biomass_min, ag_biomass_max, ag_biomass_min}{Numerich with biomass values and parameters.}

\item{seagrass_slough, seagrass_thres, seagrass_slope, time_frac}{Numerich with various parameters.}
}
\value{
void
}
\description{
Rcpp seagrass growth
}
\details{
Function to simulate processes of aboveground and belowground seagrass slough and growth.
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6
}
