% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_shuffle}
\alias{rcpp_shuffle}
\title{rcpp_shuffle}
\usage{
rcpp_shuffle(x, elements)
}
\arguments{
\item{x}{NumericVector with elements to shuffle.}

\item{elements}{Logical if vector elements or iteratos are returned.}
}
\value{
vector
}
\description{
Rcpp shuffle vector.
}
\details{
Shuffles the element of a vector (or the elements iterators).
}
\references{
How to use time-based seed taken from <http://www.cplusplus.com/reference/algorithm/shuffle/>
}
\keyword{internal}
