% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rcpp_simulate}
\alias{.rcpp_simulate}
\alias{rcpp_simulate}
\title{rcpp_simulate}
\usage{
.rcpp_simulate(
  seafloor,
  fishpop,
  nutrients_input,
  seafloor_track,
  fishpop_track,
  parameters,
  movement,
  extent,
  dimensions,
  max_i,
  min_per_i,
  save_each,
  seagrass_each,
  burn_in,
  verbose
)
}
\arguments{
\item{seafloor, fishpop}{Matrix with seafloor and fishpop data.}

\item{nutrients_input}{Vector with amount of nutrient input each time step.}

\item{seafloor_track, fishpop_track}{List with entry for each saving time step.}

\item{parameters}{List with parameters.}

\item{movement}{String specifing movement algorithm.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{min_per_i}{Integer to specify minutes per i.}

\item{save_each}{Numeric how often data should be saved to return.}

\item{seagrass_each}{Integer how often (each i * x) seagrass dynamics will be simulated.}

\item{burn_in}{Numeric with time steps used to burn in.}

\item{verbose}{Logical if TRUE, progress reports are printed.}
}
\value{
void
}
\description{
Rcpp run simulation.
}
\details{
The functions is a 'wrapper' around the following sub-processes: (i) nutrient input,
(ii) seagrass growth, (iii) detritus mineralization, (iv) movement of individuals,
(v) respiration of individuals, (vi) growth of individuals, (vii) mortality of individuals,
(viii) diffusion of nutrients/detritus, and ix) nutrient output.
}
\references{
For a detailed model description, please see Esquivel, K.E., Hesselbarth, M.H.K.,
Allgeier, J.E., In Press. Mechanistic support for increased primary production
around artificial reefs. Ecological Applications. v0.0
}
\keyword{internal}
