% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_sum}
\alias{rcpp_sum}
\title{rcpp_sum}
\usage{
rcpp_sum(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
double
}
\description{
Rcpp vector sum
}
\details{
Calculate sum of vector
}
\references{
https://teuder.github.io/rcpp4everyone_en/030_basic_usage.html
}
