% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_translate_torus}
\alias{rcpp_translate_torus}
\title{rcpp_translate_torus}
\usage{
rcpp_translate_torus(coords, extent)
}
\arguments{
\item{coords}{Add describtion}

\item{extent}{Add describtion}
}
\value{
Add describtion
}
\description{
Add describtion
}
\details{
Add describtion
}
\keyword{export}
