% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_turn_fish}
\alias{rcpp_turn_fish}
\title{rcpp_turn_fish}
\usage{
rcpp_turn_fish(fishpop, dist_values)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{dist_values}{Matrix with distance to reef values (left, straight, right).}
}
\value{
void
}
\description{
Rcpp turn fish
}
\details{
Rcpp implementation to turn fish individuals either left (-45°), straight (0°) or
right (45°) depending on which directions minimizes distance to reef.
}
\keyword{export}
