% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_update_coords}
\alias{rcpp_update_coords}
\title{rcpp_update_coords}
\usage{
rcpp_update_coords(fishpop, i, move_dist, max_dist, extent)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{i}{Integer with row id.}

\item{move_dist, max_dist}{Numeric with (maximum) movement distance.}

\item{extent}{Vector with extent of study area.}
}
\value{
void
}
\description{
Rcpp move fish population
}
\details{
Update xy coordinates and activity of fish individuals depending move_dist.
}
