% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rcpp_vec_to_map}
\alias{.rcpp_vec_to_map}
\alias{rcpp_vec_to_map}
\title{rcpp_vec_to_map}
\usage{
.rcpp_vec_to_map(key, value)
}
\arguments{
\item{key, value}{NumericVector with key and values values}
}
\value{
map
}
\description{
Rcpp matrix to map.
}
\details{
Converts to vectors to a \code{std::map} object.
}
\keyword{internal}
