% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rcpp_which}
\alias{.rcpp_which}
\alias{rcpp_which}
\title{rcpp_which}
\usage{
.rcpp_which(x, y)
}
\arguments{
\item{x}{Integer with value to find.}

\item{y}{Vector with values.}
}
\value{
vector
}
\description{
Rcpp which.
}
\details{
Returns the index of \code{x} within the \code{y} vector.
The index of the first element is 0. If element is not present within \code{y}
\code{NA} is returned. Returns only the first element if \code{x} is present
multiple times in \code{y}.
}
\keyword{internal}
