% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlognorm.R
\name{rlognorm}
\alias{rlognorm}
\title{rlognorm}
\usage{
rlognorm(n = 1, mean, sd)
}
\arguments{
\item{n}{Integer with number of generated numbers.}

\item{mean}{Numeric with mean value.}

\item{sd}{Numerich with sd value.}
}
\value{
vector
}
\description{
Random number from log-norm distribution.
}
\details{
Function to simulate random number from log-normal distribution
}
