% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{run_simulation}
\usage{
run_simulation(
  seafloor,
  fish_population,
  parameters,
  reef_attraction,
  max_i,
  min_per_i,
  save_each = 1,
  verbose = TRUE
)
}
\arguments{
\item{seafloor}{RasterBrick with environment created with \code{\link{setup_seafloor}}.}

\item{fish_population}{Data.frame population created with \code{\link{setup_fish_population}}.}

\item{parameters}{List with all model parameters.}

\item{reef_attraction}{If TRUE, individuals are attracted to AR.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{min_per_i}{Integer to specify minutes per i.}

\item{save_each}{Numeric how often data should be saved to return.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
mdl_rn
}
\description{
Core function to run model.
}
\details{
Wrapper function to run model. Executes the following sub-processes (i) simulate_seagrass
(ii) distribute_detritus (iii) simulate_movement (iv) simulate_movement (v) simulate_respiration
(vi) simulate_growth (vii) simulate_mortality and (viii) simulate_diffusion.
}
\examples{
# Add example code

}
