% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_envir_values.R
\name{setup_envir_values}
\alias{setup_envir_values}
\title{setup_envir_values}
\usage{
setup_envir_values(
  seafloor,
  ag_biomass,
  bg_biomass,
  nutrients_pool,
  detritus_pool,
  random
)
}
\arguments{
\item{seafloor}{Raster* object.}

\item{ag_biomass, bg_biomass, nutrients_pool}{Numeric with starting values.}

\item{detritus_pool}{Numeric with parameter for detritus fraction.}

\item{random}{Numeric to randomize input values by 0 = 0 percent to 1 = 100 percent.}
}
\value{
RasterBrick
}
\description{
Internal function to setup environmental values
}
\details{
Internal function to set biomass cell values of non-ARs cells.
}
