% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fish_population.R
\name{setup_fish_population}
\alias{setup_fish_population}
\title{setup_fish_population}
\usage{
setup_fish_population(
  seafloor,
  starting_values,
  parameters,
  use_log = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{seafloor}{Environment created with \code{\link{setup_seafloor}}.}

\item{starting_values}{List with all starting value parameters.}

\item{parameters}{List with all model parameters.}

\item{use_log}{Logical if TRUE, random log distribution is used.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
data.table
}
\description{
Initiate fish population.
}
\details{
Function to setup the environment.....
Center of the environment is always set to (0,0).

Parameters include ...
}
\examples{
# Add example code

}
