% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fishpop.R
\name{setup_fishpop}
\alias{setup_fishpop}
\title{setup_fishpop}
\usage{
setup_fishpop(
  seafloor,
  starting_values,
  parameters,
  use_log = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{seafloor}{Data.frame object.}

\item{starting_values}{List with all starting value parameters.}

\item{parameters}{List with all model parameters.}

\item{use_log}{Logical if TRUE, random log distribution is used.}

\item{verbose}{Logical if TRUE, information is printed.}
}
\value{
data.frame
}
\description{
Setup fish population for model run.
}
\details{
Function to setup the fish population. If \code{use_log = TRUE} the size distribution
of the fish population follows a log-norm distribution. For more information, see
\code{calc_size} (internal function). To create no fish, set \code{starting_values$pop_n = 0}.
}
\examples{
reef <- matrix(data = c(-1, 0, 0, 1, 1, 0, 0, -1, 0, 0),
ncol = 2, byrow = TRUE)

seafloor <- setup_seafloor(dimensions = c(50, 50), grain = 1,
reef = reef, starting_values = default_starting)
fishpop <- setup_fishpop(seafloor = seafloor,
starting_values = default_starting, parameters = default_parameters)

}
