% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_population.R
\name{setup_population}
\alias{setup_population}
\title{setup_population}
\usage{
setup_population(environment, starting_values, parameters, verbose = TRUE)
}
\arguments{
\item{environment}{Environment created with \code{\link{setup_environment}}.}

\item{starting_values}{List with all starting value parameters.}

\item{parameters}{List with all model parameters.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
data.table
}
\description{
Initiate fish population.
}
\details{
Function to setup the environment.....
Center of the environment is always set to (0,0).

Parameters include ...
}
\examples{
reef_matrix <- matrix(data = c(-1, 0, 0, 1, 1, 0, 0, -1, 0, 0),
ncol = 2, byrow = TRUE)

starting_values <- system.file("extdata", "starting_values.csv", package = "coRal")
parameters <- system.file("extdata", "parameters.csv", package = "coRal")
starting_values <- read_parameters(file = starting_values, sep = ";")
parameters <- read_parameters(file = parameters, sep = ";")

input_environment <- setup_environment(extent = c(100, 100), grain = 1,
reefs = reef_matrix, starting_values = starting_values, parameters = parameters)

input_population <- setup_population(environment = input_environment,
starting_values = starting_values, parameters = parameters)

}
