% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_seafloor.R
\name{setup_seafloor}
\alias{setup_seafloor}
\title{setup_seafloor}
\usage{
setup_seafloor(
  dimensions,
  grain = 1,
  reef = NULL,
  starting_values,
  random = 0,
  verbose = TRUE
)
}
\arguments{
\item{dimensions}{Vector with number of rows and columns (spatial dimensions).}

\item{grain}{Double with size of cells in x- and y-direction (spatial grain).}

\item{reef}{2-Column matrix with coordinates of artificial reefs.}

\item{starting_values}{List with all starting value parameters.}

\item{random}{Numeric to randomize input values.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
data.frame
}
\description{
Setup seafloor for model run.
}
\details{
Function to setup the seafloor. The center of the seafloor raster is always x,y (0,0).
bg_biomass and ag_biomass values are in g dry weight. nutrients_pool and deritus_pool
values are in g nutrients. If reef cells are present, the cells in the corresponding
raster layer are idntified usinge the value one.

If \code{random > 0}, the stochasticity is added to all starting values using
\code{x * (1 +- random)} as minimum and maximum values, respectively.
}
\examples{
reef <- matrix(data = c(-1, 0, 0, 1, 1, 0, 0, -1, 0, 0),
ncol = 2, byrow = TRUE)

seafloor <- setup_seafloor(dimensions = c(50, 50), grain = 1,
reef = reef, starting_values = default_starting)

}
