% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_seafloor_mdlrn.R
\name{setup_seafloor_mdlrn}
\alias{setup_seafloor_mdlrn}
\title{setup_seafloor_mdlrn}
\usage{
setup_seafloor_mdlrn(mdl_rn, verbose = TRUE, ...)
}
\arguments{
\item{mdl_rn}{\code{mdl_rn} object created with \code{\link{run_simulation}}.}

\item{verbose}{If TRUE, progress reports are printed.}

\item{...}{Additional arguments passed on to \code{\link{raster}}.}
}
\value{
RasterBrick
}
\description{
Setup seafloor from previous \code{mdl_rn} object.
}
\details{
Function to setup the environment (seafloor) from previous \code{mdl_rn} object.
Thus, the created environment will have the final values of the provided \code{mdl_rn}
object as cell values.
}
\examples{
\dontrun{
setup_seafloor_mdlrn(mdl_rn = result_rand)
}

}
