% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_diffusion.R
\name{simulate_diffusion}
\alias{simulate_diffusion}
\title{simulate_diffusion}
\usage{
simulate_diffusion(seafloor_values, cell_adj, parameters)
}
\arguments{
\item{seafloor_values}{Data.frame of seafloor values.}

\item{cell_adj}{2-column matrix with cell adjacencies.}

\item{parameters}{List with all model parameters.}
}
\value{
data.frame
}
\description{
Simulate diffusion.
}
\details{
A certain share of each cell value, specified by the diffusion parameters, is
diffused to its 8 neighboring cells.
}
