% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_fishpop_growth.R
\name{simulate_fishpop_growth}
\alias{simulate_fishpop_growth}
\title{simulate_fishpop_growth}
\usage{
simulate_fishpop_growth(
  fishpop_values,
  fishpop_track,
  n_pop,
  seafloor,
  seafloor_values,
  parameters,
  min_per_i
)
}
\arguments{
\item{fishpop_values, fishpop_track}{Matrix with fish population.}

\item{n_pop}{Numeric with number of individuals.}

\item{seafloor, seafloor_values}{RasterBrick and matrix with seafloor values.}

\item{parameters}{List with all model parameters.}

\item{min_per_i}{Integer to specify minutes per i.}
}
\value{
list
}
\description{
Simulate consumption of fish population.
}
\details{
Function to simulate consumption, possible mortality, growth and excretion of fish
population.
}
