% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_growth.R
\name{simulate_growth}
\alias{simulate_growth}
\title{simulate_growth}
\usage{
simulate_growth(
  fishpop_values,
  fishpop_track,
  pop_n,
  seafloor_values,
  parameters,
  extent,
  dimensions,
  min_per_i
)
}
\arguments{
\item{fishpop_values, fishpop_track}{Matrix with fish population.}

\item{pop_n}{Numeric with number of individuals.}

\item{seafloor_values}{RasterBrick and matrix with seafloor values.}

\item{parameters}{List with all model parameters.}

\item{extent, dimensions}{Spatial extent and dimensions of the seafloor raster}

\item{min_per_i}{Integer to specify minutes per i.}
}
\value{
list
}
\description{
Simulate consumption of fish population.
}
\details{
Function to simulate consumption, possible mortality, growth and excretion of fish
population.
}
