% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_input.R
\name{simulate_input}
\alias{simulate_input}
\title{simulate_input}
\usage{
simulate_input(seafloor_values, nutr_input, timestep)
}
\arguments{
\item{seafloor_values}{Matrix with seafloor values.}

\item{nutr_input}{Vector with amount of nutrient input each timestep.}

\item{timestep}{Integer with current timestep.}
}
\value{
matrix
}
\description{
Simulate nutrient input
}
\details{
Simulate external nutrient input to the system each timestep. The \code{nutr_input}
vector must have as many elements as \code{max_i}.
}
