% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mineralization.R
\name{simulate_mineralization}
\alias{simulate_mineralization}
\title{simulate_mineralization}
\usage{
simulate_mineralization(seafloor_values, parameters)
}
\arguments{
\item{seafloor_values}{Matrix of seafloor values.}

\item{parameters}{List with all model parameters.}
}
\value{
Matrix
}
\description{
Redistribute detritus pools
}
\details{
Function to redistribute dead detritus pool to overall detritus pool and decomposition.
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6
}
