% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mortality.R
\name{simulate_mortality}
\alias{simulate_mortality}
\title{simulate_mortality}
\usage{
simulate_mortality(
  fishpop_values,
  fishpop_track,
  pop_n,
  seafloor_values,
  parameters,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop_values, fishpop_track}{Matrix with fish population.}

\item{pop_n}{Numeric with number of individuals.}

\item{seafloor_values}{RasterBrick and matrix with seafloor values.}

\item{parameters}{List with all model parameters.}

\item{extent, dimensions}{Spatial extent and dimensions of the seafloor raster}
}
\value{
list
}
\description{
Simulate background mortality of population.
}
\details{
Function to simulate background mortality of fish population individuals.
}
