% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement.R
\name{simulate_movement}
\alias{simulate_movement}
\title{simulate_movement}
\usage{
simulate_movement(
  fishpop_values,
  pop_n,
  seafloor_values,
  reef_attraction,
  parameters,
  max_dist,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop_values}{Matrix with fish population created.}

\item{pop_n}{Numeric with number of individuals.}

\item{seafloor_values}{RasterBrick and matrix with seafloor values.}

\item{reef_attraction}{If TRUE, individuals are attracted to AR.}

\item{parameters}{List with all model parameters.}

\item{max_dist}{Numeric with maximum movement distance}

\item{extent, dimensions}{Spatial extent and dimensions of the seafloor raster}
}
\value{
matrix
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of fish population individuals.
}
