% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_output.R
\name{simulate_output}
\alias{simulate_output}
\title{simulate_output}
\usage{
simulate_output(seafloor_values, parameters)
}
\arguments{
\item{seafloor_values}{Matrix with seafloor values.}

\item{parameters}{List with all model parameters.}
}
\value{
matrix
}
\description{
Simulate nutrient output
}
\details{
Simulate loss of nutrients to to output of the system.
}
