% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seagrass.R
\name{simulate_seagrass}
\alias{simulate_seagrass}
\title{simulate_seagrass}
\usage{
simulate_seagrass(seafloor_values, parameters, cells_reef, time_frac)
}
\arguments{
\item{seafloor_values}{Matrix with seafloor values.}

\item{parameters}{List with all model parameters.}

\item{cells_reef}{Vector with cell ids of AR.}

\item{time_frac}{Integer to scale uptake from hour to min_per_i times seagrass_each.}
}
\value{
Matrix
}
\description{
Simulate seagrass.
}
\details{
Function to simulate processes of aboveground and belowground seagrass growth and
slough.
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6
}
