% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seagrass_growth.R
\name{simulate_seagrass_growth}
\alias{simulate_seagrass_growth}
\title{simulate_seagrass_growth}
\usage{
simulate_seagrass_growth(
  environment,
  starting_values,
  parameters,
  verbose = TRUE
)
}
\arguments{
\item{environment}{Environment created with \code{\link{setup_environment}}.}

\item{starting_values}{List with all starting value parameters.}

\item{parameters}{List with all model parameters.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
RasterBrick
}
\description{
Grow seagrass process.
}
\details{
Function to simulate aboveground and belowground seagrass growth.
}
