% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{split_to_sentence}
\alias{split_to_sentence}
\title{Split text into sentences}
\usage{
split_to_sentence(text)
}
\arguments{
\item{text}{input text with sentences}
}
\value{
vector of split sentences
}
\description{
\code{split_to_sentence()} splits a text into a vector of sentences
}
\details{
uses regex lookbehind to see if there is punctuation before split
}
\examples{
split_to_sentence("Okay, here's sentence 1. Here's another.")
}
