% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_mdlrn.R
\name{summarize_mdlrn}
\alias{summarize_mdlrn}
\title{summarize_mdlrn}
\usage{
summarize_mdlrn(
  result,
  what = c("seafloor", "fishpop"),
  summary = c("min", "mean", "max"),
  verbose = TRUE
)
}
\arguments{
\item{result}{mdl_rn object.}

\item{what}{Vector with 'seafloor' and/or 'fishpop' to specify what to summarize.}

\item{summary}{String with summary functions.}

\item{verbose}{Logical if TRUE, warning messages are printed.}
}
\value{
list
}
\description{
Summarize results of model run.
}
\details{
Summarize results for each time step. The \code{summary} argument
allows to specify which summary statistics are used for each cell. The selected
statistics are used by \code{\link{aggregate}} and must return a single value.

If \code{what='seafloor'}, the i) bg_biomass, ii) ag_biomass, iii) nutrients_pool,
and iv) detritus_pool are returned. If \code{what='fishpop'}, the i) length,
ii) weight, iii) died_consumption, and iv) died_background are returned.
}
\examples{
\dontrun{
summarize_mdlrn(result_rand)
}

}
