% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{summarize_results}
\alias{summarize_results}
\title{summarize_results}
\usage{
summarize_results(result, summary_fun)
}
\arguments{
\item{result}{mdl_rn object of simulation run.}

\item{summary_fun}{List with all model parameters.}
}
\value{
data.frame
}
\description{
Summarize results
}
\details{
Function to summarize results for each timestep.
}
