% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate your English to pirate speak}
\usage{
translate(english_sentence, add_random = TRUE)
}
\arguments{
\item{english_sentence}{A string with your sentence(s) in English.}

\item{add_random}{A boolean that indicates whether you want a
random pirate phrase added into your pirate speak.}
}
\value{
A string with your sentences in pirate speak.
}
\description{
\code{translate()} converts your plain English over to what a
pirate would say.
}
\examples{
## basic example code
translate("Well, hello there, sir! How can I help you?")
}
