% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_torus.R
\name{translate_torus}
\alias{translate_torus}
\title{translate_torus}
\usage{
translate_torus(coords, extent)
}
\arguments{
\item{coords}{2-Column matrix/data.frame with coordinates.}

\item{extent}{Spatial extent object of the seafloor RasterBrick.}
}
\value{
vector
}
\description{
Internal function
}
\details{
Internal function to calculate size and weight of individuals.
}
