% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_result.R
\name{animate_result}
\alias{animate_result}
\title{animate_result}
\usage{
animate_result(result, fill = "reef", ...)
}
\arguments{
\item{result}{RasterBrick with environment created with \code{\link{setup_seafloor}}.}

\item{fill}{Character to specify which values of environmental data is used as fill.}

\item{...}{Arguments passed on to \code{gganimate::animate}.}
}
\value{
data frame
}
\description{
Wrapper functions to animate results
}
\details{
Animate results
}
