% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrR-packages.R
\docType{package}
\name{arrR}
\alias{arrR}
\alias{arrR-package}
\title{arrR}
\description{
Individual-based simulation model to analyze how artificial reefs in combination
with a fish population influence seagrass growth.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Allgeier-Lab/arrR}
  \item Report bugs at \url{https://github.com/Allgeier-Lab/arrR/issues}
}

}
\author{
\strong{Maintainer}: Maximillian H.K. Hesselbarth \email{mhk.hesselbarth@gmail.com} (\href{https://orcid.org/0000-0003-1125-9918}{ORCID})

Authors:
\itemize{
  \item Kenzo Esquivel \email{kesquivel@berkeley.edu}
}

Other contributors:
\itemize{
  \item Jacob E. Allgeier \email{jeallg@umich.edu} [research team head]
}

}
