% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arrR_starting_values}
\alias{arrR_starting_values}
\title{Default starting values}
\format{
List
}
\usage{
arrR_starting_values
}
\description{
List with default starting values.
}
\references{
Citations for all parameters and starting values can be found in
Esquivel et al (2021). Mechanistic support for increased primary
production around artificial reefs. Manuscript in preparation.
}
\keyword{datasets}
