% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_dist_reefs.R
\name{int_calc_dist_reefs}
\alias{int_calc_dist_reefs}
\alias{calc_dist_reefs}
\title{calc_dist_reefs}
\usage{
int_calc_dist_reefs(object)
}
\arguments{
\item{object}{Raster* object}

\item{xy}{2-Column matrix with coordinates of artificial reefs}
}
\value{
vector
}
\description{
Internal function
}
\details{
Internal function calculate distance to reef cells.
}
\keyword{internal}
