% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_size.R
\name{calc_size}
\alias{calc_size}
\title{calc_size}
\usage{
calc_size(pop_n, pop_mean_size, pop_var_size, pop_a_grunt, pop_b_grunt)
}
\arguments{
\item{pop_n}{Numeric with starting values for number of individuals.}

\item{pop_mean_size, pop_var_size}{Numeric with parameters for mean size and variance.}

\item{pop_a_grunt, pop_b_grunt}{Numeric with parameters for weight calculation.}
}
\value{
vector
}
\description{
Internal function
}
\details{
Internal function to calculate size and weight of individuals.
}
