% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_size.R
\name{calc_size}
\alias{calc_size}
\title{calc_size}
\usage{
calc_size(
  pop_n,
  pop_mean_size,
  pop_linf_grunt,
  pop_var_size,
  pop_a_grunt,
  pop_b_grunt,
  use_log
)
}
\arguments{
\item{pop_n}{Numeric with starting values for number of individuals.}

\item{pop_mean_size, pop_linf_grunt, pop_var_size}{Numeric with parameters for mean size
and variance.}

\item{pop_a_grunt, pop_b_grunt}{Numeric with parameters for weight calculation.}

\item{use_log}{Logical if TRUE, random log distribution is used.}
}
\value{
vector
}
\description{
Internal function to calculate size
}
\details{
Internal function to calculate size and weight of fish individuals.
}
