% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_size.R
\name{int_calc_size}
\alias{int_calc_size}
\alias{calc_size}
\title{calc_size}
\usage{
int_calc_size(n, parameters)
}
\arguments{
\item{object}{Raster* object}

\item{xy}{2-Column matrix with coordinates of artificial reefs}
}
\value{
vector
}
\description{
Internal function
}
\details{
Internal function to set biomass cell values of non-ARs cells.
}
\keyword{internal}
