% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_size.R
\name{.calc_size}
\alias{.calc_size}
\alias{calc_size}
\title{calc_size}
\usage{
.calc_size(
  pop_n,
  pop_mean_size,
  pop_var_size,
  pop_linf,
  pop_a,
  pop_b,
  use_log = TRUE
)
}
\arguments{
\item{pop_n}{Numeric with starting values for number of individuals.}

\item{pop_mean_size, pop_var_size}{Numeric with parameters for mean size
and variance.}

\item{pop_linf, pop_a, pop_b}{Numeric with parameters for weight calculation.}

\item{use_log}{Logical if TRUE, random log distribution is used.}
}
\value{
list
}
\description{
Calculate dimensions of fish individuals.
}
\details{
Calculate size and weight of fish individuals based on length-weight relationships
(Froese & Pauly 2019). The starting length is randomly drawn from a log-norm or
uniform distribution.

If an uniform distribution is used, the minimum and maximum are based on 10%
and 90% of the maximum size (\code{pop_linf}), respectively.
}
\examples{
\dontrun{
calc_size(pop_n = 8, pop_mean_size = arrR_starting_values$pop_mean_size,
pop_var_size = arrR_starting_values$pop_var_size,
pop_linf = arrR_parameters$pop_linf, pop_a = arrR_parameters$pop_a,
pop_b = arrR_parameters$pop_b, use_log = TRUE)
}

}
\references{
Froese, R., Pauly, D., 2019. FishBase. World Wide Web electronic publication [WWW Document].
<www.fishbase.org>

The LENGTH-WEIGHT Table.
<https://www.fishbase.org/manual/fishbasethe_length_weight_table.htm>
}
\keyword{internal}
