% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameters.R
\name{check_parameters}
\alias{check_parameters}
\title{check_parameters}
\usage{
check_parameters(starting_values = NULL, parameters = NULL, verbose = TRUE)
}
\arguments{
\item{starting_values}{List with all starting value parameters.}

\item{parameters}{List with all model parameters.}

\item{verbose}{Logical if TRUE, progress reports are printed.}
}
\value{
void
}
\description{
Check model parameters and starting values.
}
\details{
The function checks if all required starting values and parameters are provided.
If parameters and/or starting values are missing, a vector with the name of the
corresponding values is returned.

If no arguments are provided, a list of all required parameters and values is
printed.
}
\examples{
check_parameters()

}
