% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rebirth.R
\name{create_rebirth}
\alias{create_rebirth}
\title{create_rebirth}
\usage{
create_rebirth(
  fishpop_values,
  fishpop_track,
  n_body,
  want_reserves,
  detritus_pool,
  detritus_dead,
  reason
)
}
\arguments{
\item{fishpop_values, fishpop_track}{Matrix with current fish population
and fish population of first time step.}

\item{n_body, want_reserves}{Numeric with parameters to calculate reserves.}

\item{detritus_pool, detritus_dead}{Vector with detritus values at location of
individuals.}

\item{reason}{Character specifying reason of death ('consumption' or 'background').}
}
\value{
matrix
}
\description{
Internal function to create new fish individual
}
\details{
Internal function to create new individual with same starting values as at timestep=0
if an individual dies.
}
