% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rebirth.R
\name{create_rebirth}
\alias{create_rebirth}
\title{rebirth}
\usage{
create_rebirth(
  fish_population,
  fish_population_track,
  n_body,
  want_reserves,
  detritus_pool,
  detritus_dead,
  reason
)
}
\arguments{
\item{fish_population, fish_population_track}{Data frame population created
with \code{\link{setup_fish_population}}.}

\item{n_body, want_reserves}{Numeric with parameters to calculate reserves.}

\item{detritus_pool, detritus_dead}{Vector with detritus values at location of
individual.}

\item{reason}{Character specifying reason of death ('consumption' or 'background').}
}
\value{
data.frame
}
\description{
Internal function
}
\details{
Internal function to create new individual if mortality takes place
}
