% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_parameters}
\alias{default_parameters}
\title{Default parameters}
\format{
List
}
\usage{
default_parameters
}
\description{
List with all default parameters.
}
\references{
ADD REFERENCES
}
\keyword{datasets}
