% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_dead_detritus.R
\name{distribute_dead_detritus}
\alias{distribute_dead_detritus}
\title{distribute_dead_detritus}
\usage{
distribute_dead_detritus(seafloor, detritus_death_decomp)
}
\arguments{
\item{seafloor}{Environment created with \code{\link{setup_seafloor}}.}

\item{detritus_death_decomp}{Decomposition rate of death detritus.}
}
\value{
RasterBrick
}
\description{
Redistribute dead detritus pool
}
\details{
Function to redistribute dead detritus pool to overall detritus pool
}
