% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_detritus.R
\name{distribute_detritus}
\alias{distribute_detritus}
\alias{distribute_dead_detritus}
\title{distribute_detritus}
\usage{
distribute_detritus(seafloor_values, parameters)
}
\arguments{
\item{seafloor_values}{Data.frame of seafloor values.}

\item{parameters}{List with all model parameters.}
}
\value{
RasterBrick
}
\description{
Redistribute detritus pool
}
\details{
Function to redistribute dead detritus pool to overall detritus pool and decomposition.
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6
}
