% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mdlrn.R
\name{filter_mdlrn}
\alias{filter_mdlrn}
\title{filter_mdlrn}
\usage{
filter_mdlrn(result, filter = max(result$max_i), reset = FALSE)
}
\arguments{
\item{result}{mdl_rn object of simulation run.}

\item{filter}{Vector with min/max timesteps}

\item{reset}{Logical if TRUE, cumulative seafloor values are reduced by value
before filter minimum.}
}
\value{
mdl_rn
}
\description{
Filter model run for specific timestep.
}
\details{
This functions allows to return only specific timesteps of a \code{mdl_rn} object
created with \code{\link{run_simulation}}. The function ensures that the object will
still be a \code{mdl_rn} object.
}
\examples{
 \dontrun{
filter_mdlrn(result = result_rand, filter = c(result_rand$max_i / 2, result_rand$max_i))
}

}
