% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density.R
\name{get_density}
\alias{get_density}
\title{get_density}
\usage{
get_density(result, timestep = result$max_i, normalize = FALSE)
}
\arguments{
\item{result}{mdl_rn object of simulation run.}

\item{timestep}{Integer to specify maximum timestep.}

\item{normalize}{Logical if TRUE count is divided by timesteps.}
}
\value{
data.frame
}
\description{
Get density of fish occurrence within each raster cell.
}
\details{
Calculates the fish density for each cells. This means the total count of
fish occurrence within a raster cell is divided by the maximum timestep.
}
\examples{
# Add example code

}
