% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density.R
\name{get_density}
\alias{get_density}
\title{get_density}
\usage{
get_density(result, timestep = result$max_i, plot = FALSE, base_size = 10)
}
\arguments{
\item{result}{mdl_rn object of simulation run.}

\item{timestep}{Integer to specify which timestep is plotted.}

\item{plot}{Logical if true result is plotted.}

\item{base_size}{Numeric to specify base font size.}
}
\description{
Get density of fish
}
\details{
Return raster with density of fish.
}
\examples{
\dontrun{

example code

}

}
\seealso{
\code{\link{run_simulation}}
}
