% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_limits.R
\name{get_limits}
\alias{get_limits}
\title{get_limits}
\usage{
get_limits(result, timestep = result[[1]]$max_i)
}
\arguments{
\item{result}{List with results of model run.}

\item{timestep}{Integer to specify which timestep is plotted.}
}
\value{
list
}
\description{
Get list with min and max values of seafloor for better plotting
}
\details{
Returns list with minimum and maximum values of several result runs for better
plotting. Can be provided to the \code{limits} argument of the
\code{\link{plot.mdl_rn}} function.
}
\examples{
# Add example code

}
