% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_limits.R
\name{get_limits}
\alias{get_limits}
\title{get_limits}
\usage{
get_limits(result)
}
\arguments{
\item{result}{List with results of model run.}
}
\value{
list
}
\description{
Get list with min and max values for better plotting.
}
\details{
Returns list with minimum and maximum values of several model result runs for better
plotting. The list includes the range for i) ag_biommass, ii) bg_biomass, iii)
nutrients_pool, and iv) detritus_pool. Can be provided to the \code{limits} argument
of the \code{\link{plot.mdl_rn}} function.
}
\examples{
\dontrun{
get_limits(result = list(result_rand, result_attr))
}

}
