% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbors.R
\name{get_neighbors}
\alias{get_neighbors}
\title{get_neighbors}
\usage{
get_neighbors(x, direction = 4, torus = FALSE)
}
\arguments{
\item{x}{RasterLayer.}

\item{direction}{Integer specifying if 4 or 8 neighborhood rule should be applied.}

\item{torus}{If TRUE RasterLayer is treated as torus.}
}
\value{
matrix
}
\description{
Simulate diffusion.
}
\details{
Return matrix with cell ids of neighboring cells.
}
