% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_production.R
\name{get_production}
\alias{get_production}
\title{get_production}
\usage{
get_production(result, lag = TRUE)
}
\arguments{
\item{result}{mdl_rn object.}

\item{lag}{Logical if TRUE, difference to previous time step is returned.}
}
\value{
data.frame
}
\description{
Summarize production of model run.
}
\details{
Function to summarize the bg and ag seagrass production of each time step.
The \code{lag} argument allows return the cumulative values (\code{FALSE}),
or the difference to the previous time step (\code{TRUE}).
}
\examples{
\dontrun{
get_production(result_rand)
}

}
