% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stable_values.R
\name{get_stable_values}
\alias{get_stable_values}
\title{get_stable_values}
\usage{
get_stable_values(bg_biomass, ag_biomass, parameters)
}
\arguments{
\item{bg_biomass, ag_biomass}{Numeric with starting bg and ag biomass.}

\item{parameters}{List with all model parameters.}
}
\value{
list
}
\description{
Get values for stable biomass growth.
}
\details{
Returns a list with starting values for i) the nutrients_pool and ii) the detritus_pool,
which allow stable seagrass growth if no fish individuals are present.

This means that both pools contain exactly the amount to balance the sloughed
bg_biomass and ag_biomass each time step.
}
\examples{
get_stable_values(bg_biomass = default_starting$bg_biomass,
ag_biomass = default_starting$ag_biomass, parameters = default_parameters)

}
