% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starting_values.R
\name{get_starting_values}
\alias{get_starting_values}
\title{get_starting_values}
\usage{
get_starting_values(seafloor_values, fishpop_values)
}
\arguments{
\item{seafloor_values}{Matrix with seafloor values.}

\item{fishpop_values}{Matrix population created with \code{\link{setup_fishpop}}.}
}
\value{
list
}
\description{
Get mean starting values of model run.
}
\details{
Internal function to get mean starting values during \code{\link{run_simulation}}.
Includes values for i) bg_biomass, ii) ag_biomass, iii) nutrients_pool, iv) detritus_pool,
v) pop_n, and vi) pop_mean_size. The function is needed because some of the values
can include stochasticity  and thus differ from the starting values.
}
\examples{
\dontrun{
get_starting_values(seafloor_values = seafloor_values, fishpop_values = fishpop_values)
}

}
\keyword{internal}
