% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starting_values.R
\name{get_starting_values}
\alias{get_starting_values}
\title{get_starting_values}
\usage{
get_starting_values(seafloor_values, fish_population)
}
\arguments{
\item{seafloor_values}{Data.frame with seafloor values.}

\item{fish_population}{Data.frame population created with \code{\link{setup_fish_population}}.}
}
\value{
list
}
\description{
Internal function get mean starting values
}
\details{
Internal function to get mean starting values.
}
\examples{
# Add example code

}
