% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starting_values.R
\name{get_starting_values}
\alias{get_starting_values}
\title{get_starting_values}
\usage{
get_starting_values(seafloor_values, fishpop_values)
}
\arguments{
\item{seafloor_values}{Matrix with seafloor values.}

\item{fishpop_values}{Matrix population created with \code{\link{setup_fishpop}}.}
}
\value{
list
}
\description{
Internal function get mean starting values
}
\details{
Internal function to get mean starting values.
}
\examples{
# Add example code

}
