% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_sigmoid.R
\name{int_calc_sigmoid}
\alias{int_calc_sigmoid}
\alias{int_int_calc_sigmoid}
\title{calc_sigmoid}
\usage{
int_calc_sigmoid(x, log_slope)
}
\arguments{
\item{x}{Numeric value to convert.}

\item{what}{Either 'above' or 'below'.}
}
\value{
vector
}
\description{
Calculate sigmoid to distribute uptake.
}
\details{
Calculates sigmoid that redistributes total uptake depending on x.
}
\examples{
biomass <- seq(from = 0, to = 1, by = 0.01)
f_bg <- int_calc_sigmoid(biomass, log_slope = -2)

plot(biomass, f_bg, type = "l")
lines(biomass, 1 - f_bg)

}
\keyword{internal}
