% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_seagrass_growth.R
\name{int_seagrass_growth}
\alias{int_seagrass_growth}
\title{seagrass_growth}
\usage{
int_seagrass_growth(nutrients, gamma, slough_ratio)
}
\arguments{
\item{nutrients}{Add info about parameter.}

\item{gamma}{Add info about parameter.}

\item{slough_ratio}{Add info about parameter.}

\item{growth_fraction}{Add info about parameter.}

\item{reduction_fraction}{Add info about parameter.}

\item{slough_detritus_ratio}{Add info about parameter.}
}
\value{
RasterBrick
}
\description{
Internal function to simulate above ground seagrass
}
\details{
Internal function to simulate below ground seagrass processes. This
includes i) ... ii) ...
}
\keyword{internal}
