% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_update_i.R
\name{int_update_i}
\alias{int_update_i}
\alias{int_update_i_pop}
\title{update_i_pop}
\usage{
int_update_i(data_current, data_track, ras = FALSE, increase = 1)
}
\arguments{
\item{data_current}{Data table with current population data.}

\item{data_track}{Data table with history of population data.}

\item{ras}{If TRUE data_current will be converted from raster to data.table.}

\item{increase}{Numeric how much i is increased.}
}
\value{
data.table
}
\description{
Update the time step counter.
}
\details{
Internal function to increase the time step i after each simulated time step.
}
\keyword{internal}
