% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlrn_to_raster.R
\name{mdlrn_to_raster}
\alias{mdlrn_to_raster}
\title{mdlrn_to_raster}
\usage{
mdlrn_to_raster(mdl_rn, verbose = TRUE, ...)
}
\arguments{
\item{mdl_rn}{\code{mdl_rn} object created with \code{\link{run_simulation}}.}

\item{verbose}{If TRUE, progress reports are printed.}

\item{...}{Additional arguments passed on to \code{\link{raster}}.}
}
\value{
RasterBrick
}
\description{
Convert \code{mdl_rn} object to \code{RasterBrick}.
}
\details{
Function to convert the environment (seafloor) from a previous \code{mdl_rn} object to
a \code{RasterBrick}. Thus, the created environment will have the final values of the
provided \code{mdl_rn} object as cell values. Can be used as sarting seafloor for
new simulation.
}
\examples{
\dontrun{
mdlrn_to_raster(mdl_rn = result_rand)
}

}
