% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mdl_rn.R
\name{plot.mdl_rn}
\alias{plot.mdl_rn}
\title{plot.mdl_rn}
\usage{
\method{plot}{mdl_rn}(
  x,
  what,
  summarize = FALSE,
  timestep = x$max_i,
  limits = NULL,
  base_size = 10,
  ...
)
}
\arguments{
\item{x}{mdl_rn object of simulation run.}

\item{what}{Character specifying what to plot.}

\item{summarize}{Character to specify which values of environmental data is used as fill.}

\item{timestep}{Integer to specify which timestep is plotted.}

\item{limits}{Named list with vectors with min and maximum value of values.}

\item{base_size}{Numeric to specify base font size.}

\item{...}{Not used.}
}
\description{
Plotting method for mdl_rn object
}
\details{
Plotting method for model run results simulated with \code{\link{run_simulation}}.
}
\examples{
\dontrun{

example code

}

}
\seealso{
\code{\link{run_simulation}}
}
