% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mdl_rn.R
\name{plot.mdl_rn}
\alias{plot.mdl_rn}
\title{plot.mdl_rn}
\usage{
\method{plot}{mdl_rn}(x, fill = "reef", i = x$max_i, ...)
}
\arguments{
\item{x}{Random patterns.}

\item{fill}{Character to specify which values of environmental data is used as fill.}

\item{i}{Integer to specify which timestep is plotted.}

\item{...}{Not used.}
}
\description{
Plotting method for mdl_rn object
}
\details{
Plotting method for model run results simulated with \code{\link{run_simulation}}.
}
\examples{
\dontrun{

example code

}

}
\seealso{
\code{\link{run_simulation}}
}
