% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allocation.R
\name{plot_allocation}
\alias{plot_allocation}
\title{plot_allocation}
\usage{
plot_allocation(parameters, base_size = 10)
}
\arguments{
\item{parameters}{List with model parameters.}

\item{base_size}{Numeric to specify base font size.}
}
\description{
Plotting method for seagrass growth allocation ratio.
}
\details{
Plotting method for the allocation ratio rule used during the seagrass growth
dynamics. For more information see \code{\link{rcpp_seagrass_growth}} and
\code{rcpp_allocation_ratio}.
}
\examples{
plot_allocation(parameters = arrR_parameters)

}
