% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allocation.R
\name{plot_allocation}
\alias{plot_allocation}
\title{plot_allocation}
\usage{
plot_allocation(parameters)
}
\arguments{
\item{parameters}{List with model parameters.}
}
\description{
Plotting method for seagrass growth allocation ratio.
}
\details{

}
\examples{
plot_allocation(parameters = default_parameters)

}
