% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allocation.R
\name{plot_allocation}
\alias{plot_allocation}
\title{plot_allocation}
\usage{
plot_allocation(parameters)
}
\arguments{
\item{parameters}{List with model parameters.}
}
\description{
Plotting method.
}
\details{
Plotting method for the allocation ratio rule used during the nutrient uptake
allocation within the seagrass dynamics. For more information see \code{.rcpp_seagrass_growth}
and \code{.rcpp_allocation_ratio} (internal functions).
}
\examples{
plot_allocation(parameters = arrR_parameters)

}
