% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_production.R
\name{plot_production}
\alias{plot_production}
\title{plot_production}
\usage{
plot_production(result, lag = TRUE, base_size = 10)
}
\arguments{
\item{result}{mdl_rn object of simulation run.}

\item{lag}{Logical if cumulative or difference to previous timestep should be returned.}

\item{base_size}{Numeric to specify base font size.}
}
\description{
Plotting method for production.
}
\details{
Plotting method for ag and bg production of the seagrass growth.
}
\examples{
\dontrun{
plot_production(result = result_rand)
}

}
