% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mdl_rn.R
\name{print.mdl_rn}
\alias{print.mdl_rn}
\title{print.mdl_rn}
\usage{
\method{print}{mdl_rn}(x, digits = 5, ...)
}
\arguments{
\item{x}{Random patterns.}

\item{digits}{Number of decimal places (round).}

\item{...}{Arguments passed to cat}
}
\description{
Printing method for mdl_rn object
}
\details{
Printing method for model run results simulated with \code{\link{run_simulation}}.
}
\examples{
\dontrun{

example code

}

}
\seealso{
\code{\link{run_simulation}}
}
