% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_fishpop_growth}
\alias{rcpp_calc_fishpop_growth}
\title{rcpp_calc_fishpop_growth}
\usage{
rcpp_calc_fishpop_growth(
  fishpop,
  fishpop_track,
  seafloor,
  fish_id,
  cell_id,
  pop_k_grunt,
  pop_linf_grunt,
  pop_a_grunt,
  pop_b_grunt,
  pop_n_body,
  pop_max_reserves,
  pop_want_reserves,
  min_per_i
)
}
\arguments{
\item{fishpop, fishpop_track}{Matrix with fishpop values and starting population.}

\item{seafloor}{Matrix with seafloor values.}

\item{fish_id, cell_id}{Vector with id of fish and corresponding cell ids.}

\item{pop_k_grunt, pop_linf_grunt, pop_a_grunt, pop_b_grunt}{Numeric with parameters.}

\item{pop_n_body, pop_max_reserves, pop_want_reserves, min_per_i}{Numeric with parameters.}
}
\value{
void
}
\description{
Rcpp calc growth
}
\details{
Rcpp mplementation to calculate growth of fish individuals.
}
