% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_fishpop_growth}
\alias{rcpp_calc_fishpop_growth}
\title{rcpp_calc_fishpop_growth}
\usage{
rcpp_calc_fishpop_growth(
  fishpop,
  fishpop_start,
  seafloor,
  fish_id,
  cell_id,
  growth_values,
  pop_n_body,
  pop_max_reserves,
  pop_want_reserves,
  min_per_i
)
}
\arguments{
\item{fishpop, fishpop_start}{Matrix with fishpop values.}

\item{seafloor}{Matrix with seafloor values.}

\item{fish_id, cell_id}{Vector with id of fish and corresponding cell ids.}

\item{growth_values}{Matrix with growth values.}

\item{pop_n_body, pop_max_reserves, pop_want_reserves, min_per_i}{Numeric with parameters.}
}
\value{
Matrix
}
\description{
Rcpp calc growth
}
\details{
\code{Rcpp} implementation of to calculate growth.
}
