% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_mineralization}
\alias{rcpp_calc_mineralization}
\title{rcpp_calc_mineralization}
\usage{
rcpp_calc_mineralization(
  seafloor,
  detritus_dead_decomp,
  detritus_mineralization
)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{detritus_dead_decomp, detritus_mineralization}{seafloor Numeric with parameters.}
}
\value{
void
}
\description{
Rcpp calculate mineralization
}
\details{
Rcpp implementation to calculate detritus miniralization.
}
\keyword{export}
