% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_seagrass_growth}
\alias{rcpp_calc_seagrass_growth}
\title{rcpp_calc_seagrass_growth}
\usage{
rcpp_calc_seagrass_growth(
  seafloor,
  cells_reef,
  bg_v_max,
  bg_k_m,
  bg_gamma,
  ag_v_max,
  ag_k_m,
  ag_gamma,
  bg_biomass_max,
  bg_biomass_min,
  ag_biomass_max,
  ag_biomass_min,
  seagrass_thres,
  seagrass_slope,
  detritus_ratio,
  min_per_i
)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{cells_reef}{Vector with id of reef cells.}

\item{bg_v_max, bg_k_m, bg_gamma, ag_v_max, ag_k_m, ag_gamma}{Numeric with uptake parameters.}

\item{bg_biomass_max, bg_biomass_min, ag_biomass_max, ag_biomass_min}{Numerich with biomass values and parameters.}

\item{detritus_ratio, seagrass_thres, min_per_i}{Numerich with various parameters.}
}
\value{
void
}
\description{
Rcpp calc seagrass growth
}
\details{
Rcpp implementation to calculate seagrass growth.
}
\keyword{export}
