% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calc_seagrass_growth}
\alias{rcpp_calc_seagrass_growth}
\title{rcpp_calc_seagrass_growth}
\usage{
rcpp_calc_seagrass_growth(
  seafloor,
  cells_reef,
  bg_v_max,
  bg_k_m,
  ag_v_max,
  ag_k_m,
  bg_biomass_max,
  bg_biomass_min,
  ag_biomass_max,
  ag_biomass_min,
  detritus_ratio,
  bg_thres,
  min_per_i
)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{cells_reef}{Vector with id of reef cells Add describtion}

\item{bg_v_max, bg_k_m, ag_v_max, ag_k_m}{Add describtion}

\item{bg_biomass_max, bg_biomass_min, ag_biomass_max, ag_biomass_min}{Add describtion}

\item{detritus_ratio, bg_thres, min_per_i}{Add describtion}
}
\value{
matrix
}
\description{
Rcpp calc seagrass growth
}
\details{
Internal function to calculate seagrass growth
}
\keyword{export}
