% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_cell_from_xy}
\alias{rcpp_cell_from_xy}
\title{rcpp_cell_from_xy}
\usage{
rcpp_cell_from_xy(coords, dimensions, extent)
}
\arguments{
\item{coords}{Vector with coordinates.}

\item{dimensions}{Vector with number or rows and cols}

\item{extent}{Vector with extent (xmin, xmax, ymin, ymax).}
}
\value{
int
}
\description{
Rcpp get cell id from xy
}
\details{
Rcpp implementation to get cell id from xy coordinate. Allows only one coordinate
pair at a time.
}
\references{
Code adapted from Robert J. Hijmans (2020). raster: Geographic Data Analysis
and Modeling. R package version 3.4-5. https://CRAN.R-project.org/package=raster
}
\keyword{export}
