% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_check_max_biomass}
\alias{rcpp_check_max_biomass}
\title{rcpp_check_max_biomass}
\usage{
rcpp_check_max_biomass(
  bg_biomass,
  ag_biomass,
  detritus_pool,
  bg_biomass_max,
  ag_biomass_max
)
}
\arguments{
\item{bg_biomass, ag_biomass, detritus_pool}{Numeric with values of cel..}

\item{bg_biomass_max, ag_biomass_max}{Parameters with maximum values.}
}
\value{
vector
}
\description{
Rcpp check max biomass
}
\details{
Rcpp implementation to check if current biomass values are above its max. Returns
vector with (1) bg biomass (2) ag biomass (3) detritus pool.
}
\keyword{export}
