% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_convert_nutr}
\alias{rcpp_convert_nutr}
\title{rcpp_convert_nutr}
\usage{
rcpp_convert_nutr(x, to)
}
\arguments{
\item{x}{Numeric with nutrient amount.}

\item{to}{String to specify in which unit to convert.}
}
\value{
double
}
\description{
Rcpp convert nutrients
}
\details{
Rcpp implementation to convert nutrients between g and umol.
}
\keyword{export}
