% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_create_rebirth}
\alias{rcpp_create_rebirth}
\title{rcpp_create_rebirth}
\usage{
rcpp_create_rebirth(
  fishpop,
  fishpop_start,
  seafloor,
  fish_id,
  cell_id,
  pop_n_body,
  pop_want_reserves
)
}
\arguments{
\item{fishpop, fishpop_start}{Matrix with fishpop values.}

\item{seafloor}{Matrix with seafloor values.}

\item{fish_id, cell_id}{Vector with id of fish and corresponding cell ids.}

\item{pop_n_body, pop_want_reserves}{Numeric with parameters.}
}
\value{
Matrix
}
\description{
Rcpp create new individual
}
\details{
\code{Rcpp} implementation of to calculate growth.
}
