% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_diffuse_values}
\alias{rcpp_diffuse_values}
\title{rcpp_diffuse_values}
\usage{
rcpp_diffuse_values(
  seafloor_values,
  cell_adj,
  wc_diffusion,
  detritus_diffusion,
  detritus_dead_diffusion
)
}
\arguments{
\item{x}{Vector of elements to sample from.}

\item{n}{Size of the sample.}

\item{replace}{Sample with replacement.}
}
\value{
vector
}
\description{
Rcpp sample function
}
\details{
\code{Rcpp} implementation of the \code{sample} function.
}
\seealso{
\code{\link{sample}}
}
