% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_diffuse_values}
\alias{rcpp_diffuse_values}
\title{rcpp_diffuse_values}
\usage{
rcpp_diffuse_values(
  seafloor,
  cell_adj,
  nutrients_diffusion,
  detritus_diffusion,
  detritus_fish_diffusion
)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{cell_adj}{Matrix with cell adjacencies.}

\item{nutrients_diffusion, detritus_diffusion, detritus_fish_diffusion}{Numeric with parameters.}
}
\value{
void
}
\description{
Rcpp simulate diffusion of values.
}
\details{
Simulates the diffusion of a certain share of i) nutrients_pool, ii) detritus_pool,
and iii) detritus_fish_diffusion of each cell with its neighboring cells. Scheduling
is simulated pseudo-simultaneous.
}
\keyword{internal}
