% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_diffuse_values}
\alias{rcpp_diffuse_values}
\title{rcpp_diffuse_values}
\usage{
rcpp_diffuse_values(
  seafloor,
  cell_adj,
  nutrients_diffusion,
  detritus_diffusion,
  detritus_fish_diffusion
)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{cell_adj}{Matrix with cell adjacencies.}

\item{nutrients_diffusion, detritus_diffusion, detritus_fish_diffusion}{Numeric with parameters.}
}
\value{
void
}
\description{
Rcpp diffuse values
}
\details{
Rcpp implementation to diffuse seafloor values between (queen case) neighbouring cells.
}
