% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_fishpop_growth}
\alias{rcpp_fishpop_growth}
\title{rcpp_fishpop_growth}
\usage{
rcpp_fishpop_growth(
  fishpop,
  fish_id,
  fishpop_track,
  seafloor,
  pop_k,
  pop_linf,
  pop_a,
  pop_b,
  pop_n_body,
  pop_want_reserves,
  pop_max_reserves,
  pop_consumption_prop,
  extent,
  dimensions,
  min_per_i
)
}
\arguments{
\item{fishpop, fishpop_track}{Matrix with fishpop values and starting population.}

\item{fish_id}{Vector with id of fish and corresponding cell ids.}

\item{seafloor}{Matrix with seafloor values.}

\item{pop_k, pop_linf, pop_a, pop_b}{Numeric with parameters.}

\item{pop_n_body, pop_want_reserves, pop_max_reserves, min_per_i}{Numeric with parameters.}

\item{pop_consumption_prop}{Double with consumption limit to fill reserves each timestep.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}

\item{min_per_i}{Integer to specify minutes per i.}
}
\value{
void
}
\description{
Rcpp calc growth
}
\details{
Rcpp implementation to calculate growth of fish individuals.
}
