% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_bearing}
\alias{rcpp_get_bearing}
\title{rcpp_get_bearing}
\usage{
rcpp_get_bearing(x_fish, y_fish, x_reef, y_reef)
}
\arguments{
\item{x_fish, y_fish}{Double with xy coords of fish individual.}

\item{x_reef, y_reef}{Double with xy coords of closest reef.}
}
\value{
double
}
\description{
Rcpp get bearing between two coordinates
}
\details{
Rcpp implementation to get bearing between fish individual and closest reef cell.
}
\keyword{export}
