% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rcpp_get_max_dist}
\alias{.rcpp_get_max_dist}
\alias{rcpp_get_max_dist}
\title{rcpp_get_max_dist}
\usage{
.rcpp_get_max_dist(movement, parameters, n_rand)
}
\arguments{
\item{movement}{String specifing movement algorithm.}

\item{parameters}{List with parameters.}

\item{n_rand}{Integer with amount of random numbers.}
}
\value{
double
}
\description{
Rcpp get maximum movement distance
}
\details{
Calculate double with maximum movement distance. The distance is the 95% quantile
of \code{n_rand} random numbers.

\code{movement} can be either 'rand', 'attr' or 'behav'.
}
\keyword{internal}
