% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mineralization}
\alias{rcpp_mineralization}
\title{rcpp_mineralization}
\usage{
rcpp_mineralization(seafloor, detritus_mineralization, detritus_fish_decomp)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{detritus_fish_decomp, detritus_mineralization}{seafloor Numeric with parameters.}
}
\value{
void
}
\description{
Rcpp mineralization
}
\details{
Function to redistribute fish detritus pool to overall detritus pool and decomposition.
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6
}
