% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rcpp_mineralization}
\alias{.rcpp_mineralization}
\alias{rcpp_mineralization}
\title{rcpp_mineralization}
\usage{
.rcpp_mineralization(seafloor, detritus_mineralization, detritus_fish_decomp)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{detritus_fish_decomp, detritus_mineralization}{seafloor Numeric with parameters.}
}
\value{
void
}
\description{
Rcpp simulate mineralization.
}
\details{
Simulate mineralization of the detritus pool i.e., a ratio of the detritus pool
is added to the nutrients pool. The corresponding amount is removed from the
detritus pool.

Also, simulates decomposition of the detritus fish_pool by removing a ratio and
adding it to the detritus pool.
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. <https://doi.org/10.1007/978-94-011-2342-6>
}
\keyword{internal}
