% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_modify_degree}
\alias{rcpp_modify_degree}
\title{rcpp_modify_degree}
\usage{
rcpp_modify_degree(x, y)
}
\arguments{
\item{x}{Numeric with current angle in degree.}

\item{y}{Numerich with change of degree (negative or positive).}
}
\value{
double
}
\description{
Rcpp modify degree.
}
\details{
Modify the degree of direction heading of individuals. The function ensures that
all degree are between 0 <= x <= 360.
}
\keyword{internal}
