% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mortality}
\alias{rcpp_mortality}
\title{rcpp_mortality}
\usage{
rcpp_mortality(
  fishpop,
  fishpop_track,
  seafloor,
  pop_linf,
  pop_n_body,
  pop_reserves_max,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop, fishpop_track}{Matrix with fishpop and starting fishpop values.}

\item{seafloor}{Matrix with seafloor values.}

\item{pop_linf, pop_n_body, pop_reserves_max}{Vector with parameters.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}
}
\value{
void
}
\description{
Rcpp simulate (background) mortality.
}
\details{
Function to simulate background mortality of fish individuals. The mortality
probability increases with increasing size and approximates p=1 for \code{pop_linf}.
If a individual dies, a new individual is created.
}
\keyword{internal}
