% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_behav}
\alias{rcpp_move_behav}
\title{rcpp_move_behav}
\usage{
rcpp_move_behav(
  fishpop,
  pop_reserves_thres,
  move_mean,
  move_var,
  move_reef,
  move_border,
  move_return,
  max_dist,
  coords_reef,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{pop_reserves_thres}{NumericVector with threshold of pop_reserves_max to drain prior to foraging.}

\item{move_mean, move_var}{Double with mean movement parameter.}

\item{move_reef}{Double with mean movement distance when sheltering at reef.}

\item{move_border}{Double with movement distance that surrounds reef cell border.}

\item{move_return}{Double with mean movement distance when returning to reef.}

\item{max_dist}{Maximum distance an individual can move.}

\item{coords_reef}{Matrix with ID and coords of reef cells.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}
}
\value{
void
}
\description{
Rcpp simulate movement based on bioenergetics.
}
\details{
Fish individuals move based on how much nutrients they have stored in their
reserves. There are three different movement behaviors.

If reserves above a certain threshold, individuals either shelter at reef cells (behavior 1)
or move back towards reef cells (behavior 2). If reserves are not above the threshold,
individuals move randomly across the environment to forage.
}
\keyword{internal}
