% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_behav}
\alias{rcpp_move_behav}
\title{rcpp_move_behav}
\usage{
rcpp_move_behav(
  fishpop,
  fishpop_attr,
  move_mean,
  move_sd,
  move_reef,
  move_border,
  move_return,
  max_dist,
  coords_reef,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{fishpop_attr}{Matrix with id, pop_reserves_thres_mean, and pop_reserves_consump values}

\item{move_mean, move_sd}{Double with mean movement parameter.}

\item{move_reef}{Double with mean movement distance when sheltering at reef.}

\item{move_border}{Double with movement distance that surrounds reef cell border.}

\item{move_return}{Double with mean movement distance when returning to reef.}

\item{max_dist}{Maximum distance an individual can move.}

\item{coords_reef}{Matrix with ID and coords of reef cells.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}
}
\value{
void
}
\description{
Rcpp simulate movement (behav).
}
\details{
Fish individuals move based on how much nutrients they have stored in their
reserves. There are three different movement states.

If reserves are above a certain threshold, individuals either shelter at reef cells
(state 1) or move back towards reef cells (state 2). If reserves are not above
the threshold, individuals move randomly across the seafloor to forage (state 3).
}
\keyword{internal}
