% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_fishpop}
\alias{rcpp_move_fishpop}
\title{rcpp_move_fishpop}
\usage{
rcpp_move_fishpop(
  fishpop,
  reef_dist,
  pop_thres_reserves,
  move_mean,
  move_reef,
  move_return,
  pop_visibility,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{reef_dist}{Vector with distance to reef of each cell.}

\item{pop_thres_reserves}{Vector with threshold of pop_max_reserves to drain prior to foraging}

\item{move_mean}{Numeric with parameter.}

\item{move_reef}{Double with mean movement distance when sheltering at reef}

\item{move_return}{Double with mean movement distance when returning to reef}

\item{pop_visibility}{Double with "sight" distance of fish.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}
}
\value{
void
}
\description{
Rcpp move fish population
}
\details{
Rcpp implementation to move fish individuals depending on move distance and
heading value.
"KSM": notes on code added
"Q": questions for Max
}
