% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_fishpop}
\alias{rcpp_move_fishpop}
\title{rcpp_move_fishpop}
\usage{
rcpp_move_fishpop(fishpop, move_dist, extent, pop_mean_move)
}
\arguments{
\item{fishpop}{Add describtion}

\item{move_dist}{Add describtion}

\item{extent}{Add describtion}

\item{pop_mean_move}{Add describtion}
}
\value{
Matrix
}
\description{
Rcpp move fish population
}
\details{
\code{Rcpp} implementation of to calculate growth.
}
