% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_fishpop}
\alias{rcpp_move_fishpop}
\title{rcpp_move_fishpop}
\usage{
rcpp_move_fishpop(
  fishpop,
  reef_dist,
  move_dist,
  pop_mean_move,
  pop_visibility,
  reef_attraction,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{reef_dist}{Vector with distance to reef of each cell.}

\item{move_dist}{Vector with move distance of fish individuals.}

\item{pop_mean_move}{Numeric with parameter.}

\item{pop_visibility}{Double with "sight" distance of fish.}

\item{reef_attraction}{Bool if attracted towards reef.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}
}
\value{
void
}
\description{
Rcpp move fish population
}
\details{
Rcpp implementation to move fish individuals depending on move distance and
heading value.
}
