% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_fishpop}
\alias{rcpp_move_fishpop}
\title{rcpp_move_fishpop}
\usage{
rcpp_move_fishpop(
  fishpop,
  coords_reef,
  pop_thres_reserves,
  move_border,
  move_mean,
  move_reef,
  move_return,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{coords_reef}{Matrix with coords of reef cells.}

\item{pop_thres_reserves}{Vector with threshold of pop_max_reserves to drain prior to foraging.}

\item{move_border}{Double with movement distance that surrounds reef cell border.}

\item{move_mean}{Double with mean movement parameter.}

\item{move_reef}{Double with mean movement distance when sheltering at reef.}

\item{move_return}{Double with mean movement distance when returning to reef.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}
}
\value{
void
}
\description{
Rcpp move fish population
}
\details{
Rcpp implementation to move fish individuals depending on move distance and
heading value.
"KSM": notes on code added, "MH" Some comments/ideas Max
"Q": questions for Max
}
