% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_rand}
\alias{rcpp_move_rand}
\title{rcpp_move_rand}
\usage{
rcpp_move_rand(
  fishpop,
  coords_reef,
  move_mean,
  move_var,
  move_visibility,
  max_dist,
  reef_attraction,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{coords_reef}{Matrix with coords of reef cells.}

\item{move_mean, move_var}{Double with mean and variance movement parameter.}

\item{move_visibility}{Double with "sight" distance of fish.}

\item{max_dist}{Numeric with maximum movement distance}

\item{reef_attraction}{Bool if attracted towards reef.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}
}
\value{
void
}
\description{
Rcpp move random
}
\details{
Rcpp implementation to move fish individuals randomly or attracted depending
on move distance and heading value.
}
\references{
Add reference
}
