% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_wrap}
\alias{rcpp_move_wrap}
\title{rcpp_move_wrap}
\usage{
rcpp_move_wrap(
  fishpop,
  fishpop_attr,
  movement,
  move_mean,
  move_sd,
  move_reef,
  move_border,
  move_return,
  max_dist,
  coords_reef,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{fishpop_attr}{Matrix with id and threshold of pop_reserves_max.}

\item{movement}{String specifing movement algorithm.}

\item{move_mean, move_sd}{Double with mean movement parameter.}

\item{move_reef}{Double with mean movement distance when sheltering at reef.}

\item{move_border}{Double with movement distance that surrounds reef cell border.}

\item{move_return}{Double with mean movement distance when returning to reef.}

\item{max_dist}{Maximum distance an individual can move.}

\item{coords_reef}{Matrix with ID and coords of reef cells.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}
}
\value{
void
}
\description{
Rcpp movement behavior wrapper.
}
\details{
Wrapper function around different movement algorithms. Individuals can either move
completely random (\code{'rand'}), attracted towards the artifical reef cells
\code{'attr'} or movement can be depending on the bioenergetics of the
fish individuals \code{'behav'}.
}
\keyword{internal}
