% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_nutr_output}
\alias{rcpp_nutr_output}
\title{rcpp_nutr_output}
\usage{
rcpp_nutr_output(seafloor, nutrients_output)
}
\arguments{
\item{seafloor}{Matrix with seafloor values.}

\item{nutrients_output}{Double with fraction removed from each cell.}
}
\value{
void
}
\description{
Rcpp nutrient output
}
\details{
Simulate loss of nutrients to to output of the system for each cell and timestep.
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6
}
