% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_nutr_uptake}
\alias{rcpp_nutr_uptake}
\title{rcpp_nutr_uptake}
\usage{
rcpp_nutr_uptake(nutrients, biomass, v_max, k_m, time_frac)
}
\arguments{
\item{nutrients, biomass}{Numeric with nutrient and biomass amount of cell.}

\item{v_max, k_m, time_frac}{Numeric with parameters}
}
\value{
double
}
\description{
Rcpp calculate nutrient uptake
}
\details{
Rcpp implementation to calculate nutrient uptake of cell
}
