% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_reincarnate}
\alias{rcpp_reincarnate}
\title{rcpp_reincarnate}
\usage{
rcpp_reincarnate(
  fishpop,
  fishpop_track,
  fish_id,
  seafloor,
  extent,
  dimensions,
  pop_linf,
  pop_n_body,
  pop_reserves_max,
  reason
)
}
\arguments{
\item{fishpop, fishpop_track}{Matrix with fishpop and starting fishpop values.}

\item{fish_id}{Vector with id of fish and corresponding cell ids.}

\item{seafloor}{Matrix with seafloor values.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}

\item{pop_linf, pop_n_body, pop_reserves_max}{Numeric with parameters.}

\item{reason}{String with reason of reincarnation.}
}
\value{
void
}
\description{
Rcpp reincarnate fish indivudals.
}
\details{
Creates a new individual after mortality event. The new individual has the same
value as the just died individual at the beginning of the simulation (i.e., timestep zero).
The mass difference (i.e. current mass minus mass at timestep zero) plus reserves
of the died individual are added to the detritus pool. The reincarnated individual
tries to fill its reserves from the detritus pool if enough nutrients are available.
}
