% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_rlognorm}
\alias{rcpp_rlognorm}
\title{rcpp_rlognorm}
\usage{
rcpp_rlognorm(mean, sd, min, max)
}
\arguments{
\item{mean}{Double with mean value.}

\item{sd}{Double with sd value.}

\item{min, max}{Double boundaries of random number.}
}
\value{
double
}
\description{
Rcpp rlognorm.
}
\details{
Draws random number from (truncated) log-norm distribution. Function uses log-transformed
values and a normal distribution internally.
}
\keyword{internal}
