% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_rnorm}
\alias{rcpp_rnorm}
\title{rcpp_rnorm}
\usage{
rcpp_rnorm(mean, sd, min, max)
}
\arguments{
\item{mean}{Double with mean value.}

\item{sd}{Double with sd value.}

\item{min, max}{Double boundaries of random number.}
}
\value{
double
}
\description{
Rcpp random truncated norm
}
\details{
Draws random number from (truncated) normal distribution using rejection
approach
}
\references{
https://www.cplusplus.com/reference/random/normal_distribution/?kw=normal_distribution

How to use time-based seed based on <http://www.cplusplus.com/reference/algorithm/shuffle/>
}
\keyword{internal}
