% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_shuffle}
\alias{rcpp_shuffle}
\title{rcpp_shuffle}
\usage{
rcpp_shuffle(min, max)
}
\arguments{
\item{min, max}{Integer with minum and maximum value of vector.}
}
\value{
vector
}
\description{
Rcpp shuffle
}
\details{
Get vector with random sequence from min to max.
}
