% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rcpp_shuffle}
\alias{.rcpp_shuffle}
\alias{rcpp_shuffle}
\title{rcpp_shuffle}
\usage{
.rcpp_shuffle(min, max)
}
\arguments{
\item{min, max}{Integer with min/max values.}
}
\value{
vector
}
\description{
Rcpp shuffle vector.
}
\details{
Creates vector with IDs from \code{min} to \code{max} in random order.
}
\references{
How to use time-based seed taken from
<http://www.cplusplus.com/reference/algorithm/shuffle/>
}
\keyword{internal}
