% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_translate_torus}
\alias{rcpp_translate_torus}
\title{rcpp_translate_torus}
\usage{
rcpp_translate_torus(x, y, extent)
}
\arguments{
\item{x, y}{Double with x,y coordinates}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}
}
\value{
vector
}
\description{
Rcpp translate coordinates around torus.
}
\details{
Torus translation of coordinates if they exceed the provided extent. The translation
is done until coordinate is within extent (i.e., could be translated several times
if the difference is big).
}
