% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_translate_torus}
\alias{rcpp_translate_torus}
\title{rcpp_translate_torus}
\usage{
rcpp_translate_torus(coords, extent)
}
\arguments{
\item{coords}{Vector with coordinates.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}
}
\value{
vector
}
\description{
Rcpp translate torus
}
\details{
Rcpp implementation to translate coordinates if they exceed extent.
}
\keyword{export}
