% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_turn_fish}
\alias{rcpp_turn_fish}
\title{rcpp_turn_fish}
\usage{
rcpp_turn_fish(fishpop, dist_values)
}
\arguments{
\item{fishpop}{Add describtion}

\item{dist_values}{Add describtion}
}
\value{
Add describtion
}
\description{
Add describtion
}
\details{
Add describtion
}
\keyword{export}
