% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parameters.R
\name{read_parameters}
\alias{read_parameters}
\title{read_parameters}
\usage{
read_parameters(file, return_list = TRUE, ...)
}
\arguments{
\item{file}{String with path to text file.}

\item{return_list}{Logical if true parameters are returned as list.}

\item{...}{Arguments passed on to \code{read.table}.}
}
\value{
list
}
\description{
Read parameters from text file
}
\details{
Construct a list with all default parameters. columns
}
\examples{
starting_values <- system.file("extdata", "starting_values.csv", package = "coRal")
parameters <- system.file("extdata", "parameters.csv", package = "coRal")

starting_values <- read_parameters(file = starting_values, sep = ";")
parameters <- read_parameters(file = parameters, sep = ";")

}
