% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{run_simulation}
\usage{
run_simulation(
  seafloor,
  fishpop,
  parameters,
  reef_attraction,
  max_i,
  min_per_i,
  save_each = 1,
  burn_in = 0,
  return_burnin = TRUE,
  extract = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seafloor}{RasterBrick with environment created with \code{\link{setup_seafloor}}.}

\item{fishpop}{Data.frame with fish population created with \code{\link{setup_fishpop}}.}

\item{parameters}{List with all model parameters.}

\item{reef_attraction}{If TRUE, individuals are attracted to AR.}

\item{max_i}{Integer with maximum number of simulation timesteps.}

\item{min_per_i}{Integer to specify minutes per i.}

\item{save_each}{Numeric how often data should be saved to return.}

\item{burn_in}{Numeric with timesteps used to burn in.}

\item{return_burnin}{If FALSE all timesteps < burn_in are not returned.}

\item{extract}{Character to specify if only seafloor or fishpop should be returned as data.frame}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
mdl_rn
}
\description{
Core function to run model.
}
\details{
Wrapper function to run model. Executes the following sub-processes (i) simulate_seagrass
(ii) distribute_detritus (iii) simulate_movement (iv) simulate_movement (v) simulate_respiration
(vi) simulate_growth (vii) simulate_mortality and (viii) simulate_diffusion.
}
\examples{
# Add example code

}
