% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{run_simulation}
\usage{
run_simulation(
  seafloor,
  fish_population,
  starting_values,
  parameters,
  reef_attraction,
  max_i,
  min_per_i = 120,
  verbose = TRUE
)
}
\arguments{
\item{seafloor}{RasterBrick with environment created with \code{\link{setup_seafloor}}.}

\item{fish_population}{Data frame population created with \code{\link{setup_fish_population}}.}

\item{starting_values}{List with all starting value parameters.}

\item{parameters}{List with all model parameters.}

\item{reef_attraction}{If TRUE, individuals are attracted to AR.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{min_per_i}{Integer to specify minutes per i.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
data.frame
}
\description{
Wrapper functions to run model
}
\details{
Wrapper function to run model. Executes the following sub-processes (i) ...
(ii) ...

Parameters include ...
}
\examples{
\dontrun{

run_simulation()

}

}
