% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{run_simulation}
\usage{
run_simulation(parameters, reef = NULL, max_i, verbose = FALSE)
}
\arguments{
\item{parameters}{List with all parameters.}

\item{reef}{Matrix with coordinates of reef.}

\item{max_i}{Maximum numbers of ticks to run.}

\item{verbose}{If TRUE, prints progress report.}
}
\value{
data frame
}
\description{
Wrapper functions to run model
}
\details{
Wrapper function to run model. Executes the following sub-processes (i) ...
(ii) ...

Parameters include ...
}
\examples{
\dontrun{

run_simulation()

}

}
