% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_setup_envir_values.R
\name{int_setup_envir_values}
\alias{int_setup_envir_values}
\alias{setup_envir_values}
\title{setup_envir_values}
\usage{
int_setup_envir_values(object, starting_values, parameters)
}
\arguments{
\item{object}{Raster* object}

\item{xy}{2-Column matrix with coordinates of artificial reefs}
}
\value{
vector
}
\description{
Internal function
}
\details{
Internal function to set biomass cell values of non-ARs cells.
}
\keyword{internal}
