% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_envir_values.R
\name{setup_envir_values}
\alias{setup_envir_values}
\title{setup_envir_values}
\usage{
setup_envir_values(
  seafloor,
  ag_biomass,
  bg_biomass,
  nutrients_pool,
  detritus_pool,
  random
)
}
\arguments{
\item{seafloor}{SpatRaster object.}

\item{ag_biomass, bg_biomass, nutrients_pool}{Numeric with starting values.}

\item{detritus_pool}{Numeric with parameter for detritus fraction.}

\item{random}{Numeric to randomize input values.}
}
\value{
SpatRaster
}
\description{
Setup environmental values for seafloor.
}
\details{
Create a \code{SpatRaster} with all required values for model run. The function
does not setup reef cells.

If \code{random > 0}, the stochasticity is added to all starting values using
\code{x * (1 +- random)} as minimum and maximum values, respectively.
}
\keyword{internal}
