% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_reef.R
\name{setup_reef}
\alias{setup_reef}
\title{setup_reef}
\usage{
setup_reef(seafloor, reef)
}
\arguments{
\item{seafloor}{SpatRaster object}

\item{reef}{2-Column matrix with coordinates of artificial reefs}
}
\value{
SpatRaster
}
\description{
Setup reef cells for seafloor.
}
\details{
Setup cell values of artifice reefs (reef = 1) and non-AR (reef = 0) cells.
The ag and bg biomass of all reef cells are set to \code{NA}.
}
\keyword{internal}
