% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_setup_reefs.R
\name{int_setup_reefs}
\alias{int_setup_reefs}
\alias{setup_reefs}
\title{setup_reefs}
\usage{
int_setup_reefs(object, xy)
}
\arguments{
\item{object}{Raster* object}

\item{xy}{2-Column matrix with coordinates of artificial reefs}
}
\value{
RasterBrick
}
\description{
Internal function
}
\details{
Internal function to set cell values of AR = 1 and non-AR = 0. Also, sets all
environmental values of AR to 0.
}
\keyword{internal}
