% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_reefs.R
\name{setup_reefs}
\alias{setup_reefs}
\title{setup_reefs}
\usage{
setup_reefs(object, xy, dimensions)
}
\arguments{
\item{object}{Raster* object}

\item{xy}{2-Column matrix with coordinates of artificial reefs}

\item{dimensions}{Vector with number of rows and columns (spatial dimensions).}
}
\value{
RasterBrick
}
\description{
Setup reef cells for seafloor.
}
\details{
Internal function to setup cell values of artifice reefs (reef = 1) and non-AR (reef = 0).
Sets all seafloor values of AR to NA/0. Used during \code{\link{setup_seafloor}}.
}
