% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_seafloor.R
\name{setup_seafloor}
\alias{setup_seafloor}
\title{setup_seafloor}
\usage{
setup_seafloor(
  extent,
  grain,
  reefs = NULL,
  starting_values,
  random = 0,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{extent}{Vector with number of rows and columns (spatial extent).}

\item{grain}{Vector with size of cells in x- and y-direction (spatial grain).}

\item{reefs}{2-Column matrix with coordinates of artificial reefs.}

\item{starting_values}{List with all starting value parameters.}

\item{random}{Numeric to randomize input values by 0 = 0 percent to 1 = 100 percent.}

\item{verbose}{If TRUE, progress reports are printed.}

\item{...}{Additional arguments passed on to \code{\link{raster}}.}
}
\value{
RasterBrick
}
\description{
Setup seafloor for model run.
}
\details{
Function to setup the environment (seafloor). The center of the environment is
always set to (0,0). bg_biomass and ag_biomass values are in g dry weight.
Nutrients_pool and deritus_pool values are in g nutrients. Reef cells are indicated
by \code{reef = 1}, whereas non-reef cells are indicated by \code{reef = 0}. All
other values are increased cumulative during the model run started by \code{\link{run_simulation}}.

If \code{random > 0}, the stochasticity is added to all starting values using \code{random}
as \code{x * (1 +- random)} as minimum and maximum values, respectively.
}
\examples{
reefs <- matrix(data = c(-1, 0, 0, 1, 1, 0, 0, -1, 0, 0),
ncol = 2, byrow = TRUE)

seafloor <- setup_seafloor(extent = c(100, 100), grain = 1,
reefs = reefs, starting_values = arrR_starting_values)

}
