% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_seafloor.R
\name{setup_seafloor}
\alias{setup_seafloor}
\title{setup_seafloor}
\usage{
setup_seafloor(
  extent,
  grain,
  reefs = NULL,
  starting_values,
  random = 0,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{extent}{Vector with number of rows and columns (spatial extent).}

\item{grain}{Vector with size of cells in x- and y-direction (spatial grain).}

\item{reefs}{2-Column matrix with coordinates of artificial reefs.}

\item{starting_values}{List with all starting value parameters.}

\item{random}{Numeric to randomize input values by 0 = 0 percent to 1 = 100 percent.}

\item{verbose}{If TRUE, progress reports are printed.}

\item{...}{Additional arguments passed on to \code{\link{raster}}.}
}
\value{
RasterBrick
}
\description{
Initiate environment
}
\details{
Function to setup the environment.....
Center of the environment is always set to (0,0).
All biomass values are dry values, nutrient values are g

Parameters include ...
}
\examples{
# Add example code

}
