% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_diffusion.R
\name{simulate_diffusion}
\alias{simulate_diffusion}
\title{simulate_diffusion}
\usage{
simulate_diffusion(seafloor, parameters)
}
\arguments{
\item{seafloor}{Environment created with \code{\link{setup_seafloor}}.}

\item{parameters}{List with all model parameters.}
}
\value{
RasterBrick
}
\description{
Simulate diffusion.
}
\details{
A certain share of each cell value is diffused to its 8 neighbouring cells.
}
