% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_growth.R
\name{simulate_growth}
\alias{simulate_growth}
\title{simulate_growth}
\usage{
simulate_growth(
  fishpop_values,
  fishpop_track,
  n_pop,
  seafloor,
  seafloor_values,
  parameters,
  min_per_i
)
}
\arguments{
\item{fishpop_values, fishpop_track}{Data frame population created with \code{\link{setup_fishpop_values}}.}

\item{n_pop}{Numeric with number of individuals.}

\item{seafloor, seafloor_values}{RasterBrick and matrix with seafloor values.}

\item{parameters}{List with all model parameters.}

\item{min_per_i}{Integer to specify minutes per i.}
}
\value{
list
}
\description{
Simulate consumption of fish population.
}
\details{
Function to simulate consumption, possible mortality, growth and excretion of fish
population.
}
