% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_growth.R
\name{simulate_growth}
\alias{simulate_growth}
\title{simulate_growth}
\usage{
simulate_growth(
  fish_population,
  fish_population_track,
  n_pop,
  seafloor,
  seafloor_values,
  parameters,
  min_per_i
)
}
\arguments{
\item{fish_population, fish_population_track}{Data frame population created with \code{\link{setup_fish_population}}.}

\item{n_pop}{Numeric with number of individuals.}

\item{seafloor, seafloor_values}{RasterLayer and data.frame with seafloor values.}

\item{parameters}{List with all model parameters.}

\item{min_per_i}{Integer to specify minutes per i.}
}
\value{
data.frame
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of population individuals.
}
