% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_growth.R
\name{simulate_growth}
\alias{simulate_growth}
\title{simulate_growth}
\usage{
simulate_growth(
  population,
  seafloor,
  k_grunt,
  linf_grunt,
  a_grunt,
  b_grunt,
  min_per_i,
  verbose = TRUE
)
}
\arguments{
\item{population}{Data frame population created with \code{\link{setup_population}}.}

\item{seafloor}{RasterBrick with environment created with \code{\link{setup_seafloor}}.}

\item{k_grunt, linf_grunt, a_grunt, b_grunt}{Numeric growth parameter.}

\item{min_per_i}{Integer to specify minutes per i.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
data.table
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of population individuals.
}
