% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mortality.R
\name{simulate_mortality}
\alias{simulate_mortality}
\title{simulate_mortality}
\usage{
simulate_mortality(
  fishpop_values,
  fishpop_track,
  seafloor,
  seafloor_values,
  pop_n,
  parameters,
  min_per_i
)
}
\arguments{
\item{fishpop_values, fishpop_track}{Data frame population created
with \code{\link{setup_fishpop}}.}

\item{seafloor, seafloor_values}{RasterBrick and matrix with seafloor values.}

\item{pop_n}{Numeric with number of individuals.}

\item{parameters}{List with all model parameters.}

\item{min_per_i}{Integer to specify minutes per i.}
}
\value{
list
}
\description{
Simulate background mortality of population.
}
\details{
Function to simulate background mortality of fish population individuals.
}
