% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement.R
\name{simulate_movement}
\alias{simulate_movement}
\title{simulate_movement}
\usage{
simulate_movement(
  environment,
  population,
  mean_move,
  reef_attraction,
  extent,
  verbose = TRUE
)
}
\arguments{
\item{environment}{RasterBrick with environment created with \code{\link{setup_environment}}.}

\item{population}{Data frame population created with \code{\link{setup_population}}.}

\item{mean_move}{Numeric with mean move parameter of individuals.}

\item{reef_attraction}{If TRUE, individuals are attracted to AR.}

\item{extent}{Vector with number of rows and columns (spatial extent).}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
data.table
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of population individuals.
}
