% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement.R
\name{simulate_movement}
\alias{simulate_movement}
\title{simulate_movement}
\usage{
simulate_movement(
  fish_population,
  parameters,
  extent,
  coords_reef,
  reef_attraction
)
}
\arguments{
\item{fish_population}{Data frame population created with \code{\link{setup_fish_population}}.}

\item{parameters}{List with all model parameters.}

\item{extent}{Vector with number of rows and columns (spatial extent).}

\item{coords_reef}{2-column matrix with coordinates of AR.}

\item{reef_attraction}{If TRUE, individuals are attracted to AR.}
}
\value{
data.frame
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of population individuals.
}
