% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement.R
\name{simulate_movement}
\alias{simulate_movement}
\title{simulate_movement}
\usage{
simulate_movement(
  fishpop_values,
  pop_n,
  coords_reef,
  pop_thres_reserves,
  parameters,
  extent,
  dimensions
)
}
\arguments{
\item{fishpop_values}{Matrix with fish population created.}

\item{pop_n}{Numeric with number of individuals.}

\item{coords_reef}{Matrix with coords of reef cells.}

\item{pop_thres_reserves}{Vector with proportion of max_reserves to drain prior to movement.}

\item{parameters}{List with all model parameters.}

\item{extent, dimensions}{Spatial extent and dimensions of the seafloor raster}
}
\value{
matrix
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of fish population individuals.
}
