% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement.R
\name{simulate_movement}
\alias{simulate_movement}
\title{simulate_movement}
\usage{
simulate_movement(
  fishpop_values,
  pop_n,
  seafloor_values,
  pop_thres_reserves,
  extent,
  dimensions,
  parameters
)
}
\arguments{
\item{fishpop_values}{Matrix with fish population created.}

\item{pop_n}{Numeric with number of individuals.}

\item{seafloor_values}{RasterBrick and matrix with seafloor values.}

\item{pop_thres_reserves}{Vector with proportion of max_reserves to drain prior to movement.}

\item{extent, dimensions}{Spatial extent and dimensions of the seafloor raster}

\item{parameters}{List with all model parameters.}

\item{move_reef}{move_reef Double with mean movement distance when sheltering at reef}
}
\value{
matrix
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of fish population individuals.
}
