% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement.R
\name{simulate_movement}
\alias{simulate_movement}
\title{simulate_movement}
\usage{
simulate_movement(
  fish_population,
  n_pop,
  reef_dist,
  coords_reef,
  extent,
  parameters,
  reef_attraction
)
}
\arguments{
\item{fish_population}{Data frame population created
with \code{\link{setup_fish_population}}.}

\item{n_pop}{Numeric with number of individuals.}

\item{reef_dist}{RasterLayer with distance to reef.}

\item{coords_reef}{2-column matrix with coordinates of AR.}

\item{extent}{Spatial extent object of the seafloor RasterBr.}

\item{parameters}{List with all model parameters.}

\item{reef_attraction}{If TRUE, individuals are attracted to AR.}
}
\value{
data.frame
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of population individuals.
}
