% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement.R
\name{simulate_movement}
\alias{simulate_movement}
\title{simulate_movement}
\usage{
simulate_movement(
  fish_population,
  n_pop,
  seafloor,
  seafloor_values,
  coords_reef,
  reef_attraction,
  extent,
  parameters
)
}
\arguments{
\item{fish_population}{Data frame population created
with \code{\link{setup_fish_population}}.}

\item{n_pop}{Numeric with number of individuals.}

\item{seafloor, seafloor_values}{RasterLayer and data.frame with seafloor values}

\item{coords_reef}{2-column matrix with coordinates of AR.}

\item{reef_attraction}{If TRUE, individuals are attracted to AR.}

\item{extent}{Spatial extent object of the seafloor RasterBr.}

\item{parameters}{List with all model parameters.}
}
\value{
data.frame
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of fish population individuals.
}
