% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement.R
\name{simulate_movement}
\alias{simulate_movement}
\title{simulate_movement}
\usage{
simulate_movement(
  fishpop_values,
  n_pop,
  seafloor,
  seafloor_values,
  coords_reef,
  reef_attraction,
  extent,
  parameters
)
}
\arguments{
\item{fishpop_values}{Matrix with fish population created.}

\item{n_pop}{Numeric with number of individuals.}

\item{seafloor, seafloor_values}{RasterBrick and matrix with seafloor values.}

\item{coords_reef}{2-column matrix with coordinates of AR.}

\item{reef_attraction}{If TRUE, individuals are attracted to AR.}

\item{extent}{Spatial extent object of the seafloor raster}

\item{parameters}{List with all model parameters.}
}
\value{
matrix
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of fish population individuals.
}
