% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_respiration.R
\name{simulate_respiration}
\alias{simulate_respiration}
\title{simulate_respiration}
\usage{
simulate_respiration(population, water_temp, min_per_i, verbose = TRUE)
}
\arguments{
\item{population}{Data frame population created with \code{\link{setup_population}}.}

\item{water_temp}{Numeric with water temperature.}

\item{min_per_i}{Integer to specify minutes per i.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
data.table
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of population individuals.
}
