% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_respiration.R
\name{simulate_respiration}
\alias{simulate_respiration}
\title{simulate_respiration}
\usage{
simulate_respiration(fish_population, parameters, water_temp, min_per_i)
}
\arguments{
\item{fish_population}{Data frame population created with \code{\link{setup_fish_population}}.}

\item{parameters}{List with all model parameters.}

\item{water_temp}{Numeric with water temperature.}

\item{min_per_i}{Integer to specify minutes per i.}
}
\value{
data.frame
}
\description{
Simulate movement of population.
}
\details{
Function to simulate movement of population individuals.
}
