% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seagrass.R
\name{simulate_seagrass}
\alias{simulate_seagrass}
\title{simulate_seagrass}
\usage{
simulate_seagrass(seafloor_values, parameters, cells_reef, min_per_i)
}
\arguments{
\item{seafloor_values}{Data.frame with seafloor values.}

\item{parameters}{List with all model parameters.}

\item{cells_reef}{Vector with cell ids of AR.}

\item{min_per_i}{Integer to specify minutes per i.}
}
\value{
data.frame
}
\description{
Simulate seagrass.
}
\details{
Function to simulate processes of aboveground and belowground seagrass growth and
slough.
}
\references{
DeAngelis, D.L., 1992. Dynamics of Nutrient Cycling and Food Webs. Springer
Netherlands, Dordrecht. https://doi.org/10.1007/978-94-011-2342-6
}
