% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seagrass.R
\name{simulate_seagrass}
\alias{simulate_seagrass}
\title{simulate_seagrass}
\usage{
simulate_seagrass(seafloor, parameters, min_per_i, verbose = TRUE)
}
\arguments{
\item{seafloor}{Environment created with \code{\link{setup_seafloor}}.}

\item{parameters}{List with all model parameters.}

\item{min_per_i}{Integer to specify minutes per i.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
RasterBrick
}
\description{
Simulate seagrass.
}
\details{
Function to simulate processes of above ground and below ground seagrass.
}
