% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_mdlrn.R
\name{summarize_mdlrn}
\alias{summarize_mdlrn}
\title{summarize_mdlrn}
\usage{
summarize_mdlrn(
  result,
  what = c("seafloor", "fishpop"),
  summary = c("min", "mean", "max"),
  verbose = TRUE
)
}
\arguments{
\item{result}{mdl_rn object of simulation run.}

\item{what}{Vector with 'seafloor' and/or 'fishpop' to specifiy what to summarize.}

\item{summary}{String with summary functions. Must return one value when used
with aggregate().}

\item{verbose}{Logical if warning messages should be printed.}
}
\value{
list
}
\description{
Summarize results of model run.
}
\details{
Function to summarize results for each timestep. The \code{summary} argument
allows to specify which summary statistics are used for each cell. The selected
statistics are used by \code{\link{aggregate}} and must return one value.

For the seafloor i) ag_biomass, ii) bg_biomass, iii) nutrients_pool, iv) detritus_pool
are returned. For the fish population i) length, ii) weight, iii) died_consumption, and
iv) died_background are returned.
}
\examples{
\dontrun{
summarize_mdlrn(result_rand)
}

}
