% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_mdlrn.R
\name{summarize_mdlrn}
\alias{summarize_mdlrn}
\title{summarize_mdlrn}
\usage{
summarize_mdlrn(result, summary = c("min", "mean", "max"))
}
\arguments{
\item{result}{mdl_rn object of simulation run.}

\item{summary}{String with summary functions. Must return one value when used
with aggregate().}
}
\value{
list
}
\description{
Summarize results.
}
\details{
Function to summarize results for each timestep.
}
\examples{
# Add example code

}
