% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{summarize_results}
\alias{summarize_results}
\title{summarize_results}
\usage{
summarize_results(result)
}
\arguments{
\item{result}{mdl_rn object of simulation run.}
}
\value{
data.frame
}
\description{
Summarize results
}
\details{
Function to summarize results for each timestep.
}
\examples{
# Add example code

}
